1) Bugfix of the inefficiencies in the eta and phi distributions, 2) Bugfix for total cluster energy > total TC energy, 3) 2D eta-phi corrections for cluster energies, 4) efficiency for eta and phi distributions
