+++ Makefile
@@ -1,11 +1,12 @@
 #all: emul_test-beam_Sep23 findEMax read_econt_Jul24 GenerateEmpRxFile
 
+HOME=/home/hep/idas
 CPPFLAGS=-I$(HOME)/Software/yaml-cpp/include -I common/inc -I inc -I TPGStage1Emulation/ -I TPGFEEmulation/ -I`root-config --incdir` 
 BOOST=/cvmfs/cms.cern.ch/el9_amd64_gcc12/external/boost/1.80.0-87b5de10acd2f2c8a325345ad058b814
 LDFLAGS=-L$(HOME)/Software/yaml-cpp/lib64 -L$(BOOST)/lib
 CPPFLAGSSTAGE2=-I inc/ -I TPGStage2Emulation/ -I EMPTools/ -I EMPTools/CMSSWCode -I EMPTools/HLS_arbitrary_Precision_Types/include/ -I`root-config --incdir` 
 
-all: CalcResolution.exe writeStage2MultEMPRx_TWCP.exe stage2SemiEmulator.exe compareStage2TWCP_FWvsEmul_allBxs.exe stage2TWCPEmulEMP.exe stage2SemiEmulationTCBitsAccumulatorFW.exe stage2SemiEmulationTCBits.exe stage2HtoTauTauEnergyCorrelation.exe testfp.exe stage2CPCalc.exe #compareCPPyCpp.exe stage2CPCalc.exe readCPPyEmulOut.exe writeStage2EMPRx_TWCP.exe writeStage2MultEMPRx_TWCP.exe compareStage2TWCP_FWvsEmul.exe compareStage2TWCP_FWvsEmul_allBxs.exe stage2TWCPEmulEMP.exe  #compareStage2TowerFWvsEmul.exe stage2TowerEmulEMP.exe readlpGBTpairEvents.exe  #checkStage2Config.exe testStage1Stage2Emulation.exe  writeStage2EMPRx.exe EmulTowerPRRTest.exe #readNTuple.exe  fillInputData.exe testStage2SemiClustering.exe stage2HtoTauTauEnergyCorrelation.exe ntupleMCInfo.exe vbfjet.exe  TowerPreEmulTest.exe  TestUnpackerTCProcInterface.exe #findFixedpattern.exe findEMax.exe tpgdata_3T_tcproc.exe  tpgdata_3T_fe.exe  scanadc_Sep24.exe dump_event.exe tpgdata_2T_fe.exe emul_Sep24.exe emul_3T_Sep24.exe  #loop_emul_Sep24.exe scanconfigval_Sep24.exe emul_Sep24.exe validateFixedADC.exe # findEMax.exe GenerateEmpRxFile.exe dump_event.exe emul_test-beam_Sep23.exe 
+all: CalcResolution.exe stage2SemiEmulator.exe #writeStage2MultEMPRx_TWCP.exe  compareStage2TWCP_FWvsEmul_allBxs.exe stage2TWCPEmulEMP.exe stage2SemiEmulationTCBitsAccumulatorFW.exe stage2SemiEmulationTCBits.exe stage2HtoTauTauEnergyCorrelation.exe testfp.exe stage2CPCalc.exe #compareCPPyCpp.exe stage2CPCalc.exe readCPPyEmulOut.exe writeStage2EMPRx_TWCP.exe writeStage2MultEMPRx_TWCP.exe compareStage2TWCP_FWvsEmul.exe compareStage2TWCP_FWvsEmul_allBxs.exe stage2TWCPEmulEMP.exe  #compareStage2TowerFWvsEmul.exe stage2TowerEmulEMP.exe readlpGBTpairEvents.exe  #checkStage2Config.exe testStage1Stage2Emulation.exe  writeStage2EMPRx.exe EmulTowerPRRTest.exe #readNTuple.exe  fillInputData.exe testStage2SemiClustering.exe stage2HtoTauTauEnergyCorrelation.exe ntupleMCInfo.exe vbfjet.exe  TowerPreEmulTest.exe  TestUnpackerTCProcInterface.exe #findFixedpattern.exe findEMax.exe tpgdata_3T_tcproc.exe  tpgdata_3T_fe.exe  scanadc_Sep24.exe dump_event.exe tpgdata_2T_fe.exe emul_Sep24.exe emul_3T_Sep24.exe  #loop_emul_Sep24.exe scanconfigval_Sep24.exe emul_Sep24.exe validateFixedADC.exe # findEMax.exe GenerateEmpRxFile.exe dump_event.exe emul_test-beam_Sep23.exe 
 
 emul_test-beam_Sep23.exe:  test-beam_Sep23_macros/emul_test-beam_Sep23.cpp inc/*.*  TPGFEEmulation/*.hh TPGStage1Emulation/*.hh common/inc/*.h
 	g++ $(LDFLAGS) $(CPPFLAGS) test-beam_Sep23_macros/emul_test-beam_Sep23.cpp  -l yaml-cpp `root-config --libs --cflags` -o emul_test-beam_Sep23.exe
