+++ test-stage2_Nov24/CalcResolutionVBF.cpp
@@ -125,8 +125,6 @@ int main(int argc, char** argv)
     pt_clusThresh = 150.; 
     pt_clusThresh_etaeff = 200.; 
   }else if(sampleType.find("singlePion")!=std::string::npos){
-    // pt_clusThresh = 150.; 
-    // pt_clusThresh_etaeff = 200.; 
     pt_clusThresh = 100.; 
     pt_clusThresh_etaeff = 150.; 
   }
@@ -253,11 +251,7 @@ int main(int argc, char** argv)
   std::vector<float>  *tc_pt = 0 ;
   tr->SetBranchStatus("tc_pt",1);
   tr->SetBranchAddress("tc_pt" , &tc_pt);
-
-  std::vector<float>  *tc_mipPt = 0 ;
-  tr->SetBranchStatus("tc_mipPt",1);
-  tr->SetBranchAddress("tc_mipPt" , &tc_mipPt);
-
+  
   std::vector<float>  *tc_eta = 0 ;
   tr->SetBranchStatus("tc_eta",1);
   tr->SetBranchAddress("tc_eta" , &tc_eta);
@@ -359,16 +353,13 @@ int main(int argc, char** argv)
   
   //Eta : outermost(Sci) = 1.321, innermost(Si) =  3.152
   //use delta-Eta = 0.0873 (or mutltiple) and make range for |\eta|
-
-  // float deltaEta = 0.0873;
-  // const int nJetEtaBins = 21;
-  // Float_t jetEtaBin[22];
-  // jetEtaBin[0] = 1.321;
-  // for(int ieta=1;ieta<=nJetEtaBins;ieta++)
-  //   jetEtaBin[ieta] = jetEtaBin[0] + ieta*deltaEta;
-  
-  const int nJetEtaBins = 6;
-  Float_t jetEtaBin[7] = {1.321, 1.7, 2.0, 2.3, 2.6, 2.9, 3.152} ;
+  float deltaEta = 0.0873;
+  const int nJetEtaBins = 21;
+  Float_t jetEtaBin[22];
+  jetEtaBin[0] = 1.321;
+  for(int ieta=1;ieta<=nJetEtaBins;ieta++)
+    jetEtaBin[ieta] = jetEtaBin[0] + ieta*deltaEta;
+  //jetEtaBin[21] would be 3.1543;
   
   const int nJetPtBins = 43;
   Float_t jetPtBin[44] = {0, 15., 16., 18., 20., 22., 25., 30., 35., 40., 45., 50., 55., 60., 65., 70., 75., 80., 85., 90., 95., 100.,  
@@ -379,8 +370,7 @@ int main(int argc, char** argv)
   
   int nPtResoBins = 400;
   int nPtRatioBins = 1000;
-  float etaMinCore = 1.7, etaMaxCore = 2.9;
-  //float etaMinCore = 1.6, etaMaxCore = 3.0;
+  
   //////////////// 2D pt,eta histos ////////////////////////////
   TH1F ***hGenClusPtReso, ***hGenTCPtReso, ***hTCClusPtReso;  
   hGenClusPtReso = new TH1F**[nJetEtaBins];
@@ -393,13 +383,13 @@ int main(int argc, char** argv)
     for(int ipt=0;ipt<nJetPtBins;ipt++){
       hGenClusPtReso[ieta][ipt] = new TH1F(Form("hGenClusPtReso_%d_%d",ieta,ipt),
 					   Form("hGenClusPtReso #eta:(%4.3f-%4.3f) p_{T}:(%2.0f-%2.0f) GeV",jetEtaBin[ieta],jetEtaBin[ieta+1],jetPtBin[ipt],jetPtBin[ipt+1]),
-					   nPtResoBins,0.,40.);
+					   nPtResoBins,-400.,400.);
       hGenTCPtReso[ieta][ipt] = new TH1F(Form("hGenTCPtReso_%d_%d",ieta,ipt),
 					   Form("hGenTCPtReso #eta:(%4.3f-%4.3f) p_{T}:(%2.0f-%2.0f) GeV",jetEtaBin[ieta],jetEtaBin[ieta+1],jetPtBin[ipt],jetPtBin[ipt+1]),
-					   nPtResoBins,0.,40.);
+					   nPtResoBins,-400.,400.);
       hTCClusPtReso[ieta][ipt] = new TH1F(Form("hTCClusPtReso_%d_%d",ieta,ipt),
 					  Form("hTCClusPtReso #eta:(%4.3f-%4.3f) p_{T}:(%2.0f-%2.0f) GeV",jetEtaBin[ieta],jetEtaBin[ieta+1],jetPtBin[ipt],jetPtBin[ipt+1]),
-					  nPtResoBins,0.,40.);
+					  nPtResoBins,-400.,400.);
     }// jet et
   }//jet eta
   //////////////////////////////////////////////////////////
@@ -488,6 +478,15 @@ int main(int argc, char** argv)
   TH2D *h2TCVsClusPt = new TH2D("h2TCVsClusPt","h2TCVsClusPt", 100, 0., 500., 100, 0., 500.);
   ////////////////////////////////////////////////////////
   
+  //////////////// 1D pt //////////////////////////////////
+  TH1F **hGenTCPtResovsTCPt1D;
+  hGenTCPtResovsTCPt1D = new TH1F*[100];
+  for(int ipt=0;ipt<100;ipt++){
+    hGenTCPtResovsTCPt1D[ipt] = new TH1F(Form("hGenTCPtResovsTCPt1D_%d",ipt),
+					 Form("hGenTCPtResovsTCPt1D p_{T}:(%2.0f-%2.0f) GeV",float(ipt),float(ipt+1)),
+					 nPtResoBins,-400.,400.);
+  }
+  //////////////// 1D pt //////////////////////////////////
   
   //////////////// 1D pt log x//////////////////////////////////
   double xm1 = -3;
@@ -506,35 +505,33 @@ int main(int argc, char** argv)
     }
   }
   
+  TH1F **hGenTCPtResovsTCPt1DLog;
+  hGenTCPtResovsTCPt1DLog = new TH1F*[int(logbins.size()-1)];
+  for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
+    hGenTCPtResovsTCPt1DLog[ipt] = new TH1F(Form("hGenTCPtResovsTCPt1DLog_%d",ipt),
+					    Form("hGenTCPtResovsTCPt1DLog p_{T}:(%2.1f-%2.1f) MeV",logbins.at(ipt),logbins.at(ipt+1)),
+					    nPtResoBins,-400.,400.);
+  }
+  
   TH1F **hGenjetByTCPtSumVsTCPt1DLog;
   hGenjetByTCPtSumVsTCPt1DLog = new TH1F*[int(logbins.size()-1)];
   for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
     hGenjetByTCPtSumVsTCPt1DLog[ipt] = new TH1F(Form("hGenjetByTCPtSumVsTCPt1DLog_%d",ipt),
-						Form("hGenjetByTCPtSumVsTCPt1DLog p_{T}:(%2.1f-%2.1f) MeV",logbins.at(ipt),logbins.at(ipt+1)),
-						nPtResoBins,0.,40.);
+					    Form("hGenjetByTCPtSumVsTCPt1DLog p_{T}:(%2.1f-%2.1f) MeV",logbins.at(ipt),logbins.at(ipt+1)),
+					    nPtResoBins,0.,4.);
   }
   //////////////// 1D pt //////////////////////////////////
 
   //////////////// 2D pt //////////////////////////////////
   const int nTCPtBins = 100;
-  TH1F ***hGenjetByTCPtSumVsTCPt2DDetLog;
-  hGenjetByTCPtSumVsTCPt2DDetLog = new TH1F**[7];
+  TH1F ***hGenTCPtResovsTCPt2D;
+  hGenTCPtResovsTCPt2D = new TH1F**[7];
   for(int idet=0;idet<7;idet++){
-    hGenjetByTCPtSumVsTCPt2DDetLog[idet] = new TH1F*[int(logbins.size()-1)];
-    for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
-      hGenjetByTCPtSumVsTCPt2DDetLog[idet][ipt] = new TH1F(Form("hGenjetByTCPtSumVsTCPt2DDetLog_%d_%d",idet,ipt),
-						 Form("hGenjetByTCPtSumVsTCPt2DDetLog idet: %d, p_{T}:(%2.0f-%2.0f) GeV",idet,logbins.at(ipt),logbins.at(ipt+1)),
-						 nPtResoBins,0.,40.);
-    }
-  }
-  TH1F ***hGenjetByTCPtSumVsTCPt2DLog;
-  hGenjetByTCPtSumVsTCPt2DLog = new TH1F**[nJetEtaBins];
-  for(int ieta=0;ieta<nJetEtaBins;ieta++){
-    hGenjetByTCPtSumVsTCPt2DLog[ieta] = new TH1F*[int(logbins.size()-1)];
-    for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
-      hGenjetByTCPtSumVsTCPt2DLog[ieta][ipt] = new TH1F(Form("hGenjetByTCPtSumVsTCPt2DLog_%d_%d",ieta,ipt),
-							Form("hGenjetByTCPtSumVsTCPt2DLog #eta:(%4.3f-%4.3f), p_{T}:(%2.0f-%2.0f) GeV",jetEtaBin[ieta],jetEtaBin[ieta+1],logbins.at(ipt),logbins.at(ipt+1)),
-							nPtResoBins,0.,40.);
+    hGenTCPtResovsTCPt2D[idet] = new TH1F*[nTCPtBins];
+    for(int ipt=0;ipt<nTCPtBins;ipt++){
+      hGenTCPtResovsTCPt2D[idet][ipt] = new TH1F(Form("hGenTCPtResovsTCPt2D_%d_%d",idet,ipt),
+						 Form("hGenTCPtResovsTCPt2D idet: %d, p_{T}:(%2.0f-%2.0f) GeV",idet,float(ipt),float(ipt+1)),
+						 nPtResoBins,-400.,400.);
     }
   }
   //////////////// 1D pt //////////////////////////////////
@@ -569,7 +566,7 @@ int main(int argc, char** argv)
   //   }// jet et
   // }//jet eta
   // ////////////////////////////////////////////////////////
-  
+
   // /////////////////// 3D CEH only correction /////////////////
   // TH1F ****hHadGenClusPtReso3D;
   // hHadGenClusPtReso3D = new TH1F***[7];
@@ -585,22 +582,8 @@ int main(int argc, char** argv)
   //   }//jet eta
   // }//idet loop
   // ////////////////////////////////////////////////////////
-  
+
   //////////////// 3D pt,eta histos ////////////////////////////
-  // TH1F ****hGenjetByTCPtSumVsTCPt3DLog;
-  // hGenjetByTCPtSumVsTCPt3DLog = new TH1F***[7];
-  // for(int idet=0;idet<7;idet++){
-  //   hGenjetByTCPtSumVsTCPt3DLog[idet] = new TH1F**[nJetEtaBins];
-  //   for(int ieta=0;ieta<nJetEtaBins;ieta++){
-  //     hGenjetByTCPtSumVsTCPt3DLog[idet][ieta] = new TH1F*[int(logbins.size()-1)];
-  //     for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
-  // 	hGenjetByTCPtSumVsTCPt3DLog[idet][ieta][ipt] = new TH1F(Form("hGenjetByTCPtSumVsTCPt3DLog_%d_%d_%d",idet,ieta,ipt),
-  // 								Form("hGenjetByTCPtSumVsTCPt3DLog idet: %d, #eta:(%4.3f-%4.3f) , p_{T}:(%2.0f-%2.0f) GeV",idet,jetEtaBin[ieta],jetEtaBin[ieta+1],logbins.at(ipt),logbins.at(ipt+1)),
-  // 								nPtResoBins,0.,40.);
-  //     }
-  //   }
-  // }
-  
   TH1F ****hGenClusPtReso3D; 
   hGenClusPtReso3D = new TH1F***[7];
   for(int idet=0;idet<7;idet++){
@@ -610,12 +593,12 @@ int main(int argc, char** argv)
       for(int ipt=0;ipt<nJetPtBins;ipt++){
 	hGenClusPtReso3D[idet][ieta][ipt] = new TH1F(Form("hGenClusPtReso3D_%d_%d_%d",idet,ieta,ipt),
 						 Form("hGenClusPtReso3D det:%d, #eta:(%4.3f-%4.3f) p_{T}:(%2.0f-%2.0f) GeV",idet,jetEtaBin[ieta],jetEtaBin[ieta+1],jetPtBin[ipt],jetPtBin[ipt+1]),
-						 nPtResoBins,0.,40.);
+						 nPtResoBins,-400.,400.);
       }// jet et
     }//jet eta
   }//idet loop
   //////////////////////////////////////////////////////////
-  
+
   ///////////Energy correction for Cl3d cluster //////////
   TH1F **hGenClusPtReso1Dcl3d, **hNewOldPtReso1Dcl3d;
   hGenClusPtReso1Dcl3d = new TH1F*[nJetPtBins];
@@ -631,57 +614,29 @@ int main(int argc, char** argv)
   TH2D *h2GenVsClusPtcl3d = new TH2D("h2GenVsClusPtcl3d","h2GenVsClusPtcl3d", 100, 0., 500., 100, 0., 500.);
   TH2D *h2NewVsOldPtcl3d = new TH2D("h2NewVsOldPtcl3d","h2NewVsOldPtcl3d", 100, 0., 500., 100, 0., 500.);
   TProfile *pNewVsOldPtcl3d = new TProfile("pNewVsOldPtcl3d","pNewVsOldPtcl3d", 100, 0., 500., 0., 500.);
-  ////////////////////////////////////////////////////////  
+  ////////////////////////////////////////////////////////
+  
   TH2F *hJetEtaPtBin = new TH2F("hJetEtaPtBin","hJetEtaPtBin", nJetEtaBins, jetEtaBin, nJetPtBins, jetPtBin);
   TH2F *hJetEtaPtBinDet = new TH2F("hJetEtaPtBinDet","hJetEtaPtBinDEt", nJetEtaBins, jetEtaBin, nJetPtBins, jetPtBin);
   
-  TH2F *hJetEtaTCPtBin = new TH2F("hJetEtaTCPtBin","hJetEtaTCPtBin", nJetEtaBins, jetEtaBin[0], jetEtaBin[nJetEtaBins-1], 100, 0.,100.);
+  TH2F *hJetEtaTCPtBin = new TH2F("hJetEtaTCPtBin","hJetEtaTCPtBin", 21, 1.6, 3.0, 100, 0.,100.);
   const Double_t *plogbins = logbins.data();
-  TH2F *hJetEtaTCPtBinLog = new TH2F("hJetEtaTCPtBinLog","hJetEtaTCPtBinLog", nJetEtaBins, jetEtaBin[0], jetEtaBin[nJetEtaBins-1], int(logbins.size()-1), plogbins);
+  TH2F *hJetEtaTCPtBinLog = new TH2F("hJetEtaTCPtBinLog","hJetEtaTCPtBinLog", 21, 1.6, 3.0, int(logbins.size()-1), plogbins);
   
-  ////////////// TC-vs-genpt /////////////////////////
-  TH2D *h2TCvsGenPt = new TH2D("h2TCvsGenPt","h2TCvsGenPt", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TCvsGenPt_CoreEta = new TH2D("h2TCvsGenPt_CoreEta","h2TCvsGenPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TCvsGenPt_CoreEta_rozcut0p15 = new TH2D("h2TCvsGenPt_CoreEta_rozcut0p15","h2TCvsGenPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TCvsGenPt_CoreEta_mpptcut2 = new TH2D("h2TCvsGenPt_CoreEta_mpptcut2","h2TCvsGenPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  
-  TH2D *h2TCvsGenPt_CEE = new TH2D("h2TCvsGenPt_CEE","h2TCvsGenPt_CEE", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TCvsGenPt_CEE_CoreEta = new TH2D("h2TCvsGenPt_CEE_CoreEta","h2TCvsGenPt_CEE_CoreEta", 250, 0., 500., 250, 0., 500.);
-  
-  TProfile *pTCvsGenPt = new TProfile("pTCvsGenPt","pTCvsGenPt", 250, 0., 500., 0., 500.);
-  TProfile *pTCvsGenPt_CoreEta = new TProfile("pTCvsGenPt_CoreEta","pTCvsGenPt_CoreEta", 250, 0., 500., 0., 500.);
-  TProfile *pTCvsGenPt_CoreEta_rozcut0p15 = new TProfile("pTCvsGenPt_CoreEta_rozcut0p15","pTCvsGenPt_CoreEta_rozcut0p15", 250, 0., 500., 0., 500.);
-  TProfile *pTCvsGenPt_CoreEta_mpptcut2 = new TProfile("pTCvsGenPt_CoreEta__mpptcut2","pTCvsGenPt_CoreEta__mpptcut2", 250, 0., 500., 0., 500.);
-  TProfile *pTCvsGenPt_CEE = new TProfile("pTCvsGenPt_CEE","pTCvsGenPt_CEE", 250, 0., 500., 0., 500.);
-  TProfile *pTCvsGenPt_CEE_CoreEta = new TProfile("pTCvsGenPt_CEE_CoreEta","pTCvsGenPt_CEE_CoreEta", 250, 0., 500., 0., 500.);
-  
-  ////////////// genpt/totTC-vs-(pt/eta) /////////////////////////
-  TH2D *h2GenjetTCPtsumvsGenPt = new TH2D("h2GenjetTCPtsumvsGenPt","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 200, 0., 10.);
-  TH2D *h2GenjetTCPtsumvsGenEta = new TH2D("h2GenjetTCPtsumvsGenEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #eta^{genJet}",  90, 0., 3.93, 200, 0., 10.);
-  TProfile *pGenjetTCPtsumvsGenPt = new TProfile("pGenjetTCPtsumvsGenPt","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 0., 10.);
-  TProfile *pGenjetTCPtsumvsGenEta = new TProfile("pGenjetTCPtsumvsGenEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #eta^{genJet}", 90, 0., 3.93, 0., 10.);
+  TH1D *hMaxLayerE = new TH1D("hMaxLayerE","Layer with maximum energy",50, -0.5, 49.5);
+  TH1D *hMaxDetE = new TH1D("hMaxDetE","dettype with maximum energy",10, -0.5, 9.5);
 
-  TH2D *h2GenjetTCPtsumvsGenPt_CoreEta = new TH2D("h2GenjetTCPtsumvsGenPt_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 200, 0., 10.);
-  TProfile *pGenjetTCPtsumvsGenPt_CoreEta = new TProfile("pGenjetTCPtsumvsGenPt_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 0., 10.);
-
-  TH2D *h2GenjetTCPtsumvsGenPt_CEE = new TH2D("h2GenjetTCPtsumvsGenPt_CEE","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 200, 0., 10.);
-  TH2D *h2GenjetTCPtsumvsGenEta_CEE = new TH2D("h2GenjetTCPtsumvsGenEta_CEE","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #eta^{genJet}", 90, 0., 3.93, 200, 0., 10.);
-  TH2D *h2GenjetTCPtsumvsGenPt_CEE_CoreEta = new TH2D("h2GenjetTCPtsumvsGenPt_CEE_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 200, 0., 10.);
-  TH2D *h2GenjetTCPtsumvsTCPtsum_CEE_CoreEta = new TH2D("h2GenjetTCPtsumvsTCPtsum_CEE_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #Sigmap_{T}^{TC}", 100, 0., 500., 200, 0., 10.);
-  
-  TProfile *pGenjetTCPtsumvsGenPt_CEE = new TProfile("pGenjetTCPtsumvsGenPt_CEE","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 0., 10.);
-  TProfile *pGenjetTCPtsumvsGenEta_CEE = new TProfile("pGenjetTCPtsumvsGenEta_CEE","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #eta^{genJet}", 90, 0., 3.93, 0., 10.);
-  TProfile *pGenjetTCPtsumvsGenPt_CEE_CoreEta = new TProfile("pGenjetTCPtsumvsGenPt_CEE_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs p_{T}^{genJet}", 100, 0., 500., 0., 10.);
-  TProfile *pGenjetTCPtsumvsTCPtsum_CEE_CoreEta = new TProfile("pGenjetTCPtsumvsTCPtsum_CEE_CoreEta","p_{T}^{genJet}/#Sigmap_{T}^{TC} vs #Sigmap_{T}^{TC}", 100, 0., 500., 0., 10.);
-
-  /////////////////// clus-vs-tc ////////////////////////////////
-  TH2D *h2ClusVsTCPt = new TH2D("h2ClusVsTCPt","h2ClusVsTCPt", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2ClusVsTCPt_CoreEta = new TH2D("h2ClusVsTCPt_CoreEta","h2ClusVsTCPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2ClusVsTCPt_CoreEta_rozcut0p15 = new TH2D("h2ClusVsTCPt_CoreEta_rozcut0p15","h2ClusVsTCPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TotClusVsTCPt = new TH2D("h2TotClusVsTCPt","h2TotClusVsTCPt", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TotClusVsTCPt_CoreEta = new TH2D("h2TotClusVsTCPt_CoreEta","h2TotClusVsTCPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-  TH2D *h2TotClusVsTCPt_CoreEta_rozcut0p15 = new TH2D("h2TotClusVsTCPt_CoreEta_rozcut0p15","h2TotClusVsTCPt_CoreEta", 250, 0., 500., 250, 0., 500.);
-
+  ////////////// Check Eloss in CEE /////////////////////////
+  TH2D *h2TCvsGenPt = new TH2D("h2TCvsGenPt","h2TCvsGenPt", 100, 0., 100., 200, 0., 400.);
+  TProfile *pTCvsGenPt = new TProfile("pTCvsGenPt","pTCvsGenPt", 100, 0., 100., 0., 400.);
+  TH2D *h2GenvsTCPt = new TH2D("h2GenvsTCPt","h2GenvsTCPt", 200, 0., 400., 100, 0., 100.);
+  TProfile *pGenvsTCPt = new TProfile("pGenvsTCPt","pGenvsTCPt", 100, 0., 400., 0., 100.);
+  TH2D *h2TCvsGenPtDet[7];
+  TProfile *pTCvsGenPtDet[7];
+  for(int idet=0;idet<7;idet++){
+    h2TCvsGenPtDet[idet] = new TH2D(Form("h2TCvsGenPtDet_%d",idet),Form("h2TCvsGenPtDet_%d",idet), 100, 0., 100., 200, 0., 400.);
+    pTCvsGenPtDet[idet] = new TProfile(Form("pTCvsGenPtDet_%d",idet),Form("pTCvsGenPtDet_%d",idet), 100, 0., 100., 0., 400.);
+  }
   ///////////////////////////////////////////////////////////
   TH2D *h2TCPtvsGenPtByTCPtsum = new TH2D("h2TCPtvsGenPtByTCPtsum","h2TCPtvsGenPtByTCPtsum", 6000, 0., 60000., 200, -4., 4.); //TCpt in MeV
   TProfile *pTCPtvsGenPtByTCPtsum = new TProfile("pTCPtvsGenPtByTCPtsum","pTCPtvsGenPtByTCPtsum", 6000, 0., 60000., -4., 4.);
@@ -690,54 +645,16 @@ int main(int argc, char** argv)
   for(int idet=0;idet<7;idet++){
     pTCPtvsGenPtByTCPtsumDet[idet] = new TProfile(Form("pTCPtvsGenPtByTCPtsumDet_%d",idet),Form("pTCPtvsGenPtByTCPtsumDet_%d",idet), 6000, 0., 60000., -4., 4.);
   }
-  ////////////////////Unscaled Pt,eta, phi of TCs//////////////////////////////
-  TH1D *hTCPt = new TH1D("hTCPt","p_{T} distribution of TCs",500,0.0,500.0);
-  hTCPt->GetXaxis()->SetTitle("p_{T} (GeV)");
-  hTCPt->GetYaxis()->SetTitle("Entries");
-  TH1D *hTCEta = new TH1D("hTCEta","#eta distribution of TCs",1000,-4.15,4.15);
-  hTCEta->GetXaxis()->SetTitle("#eta");
-  hTCEta->GetYaxis()->SetTitle("Entries");
-  TH1D *hTCPhiPos = new TH1D("hTCPhiPos","#phi distribution of TCs for #eta>0",400,-200.,200.);
-  hTCPhiPos->GetXaxis()->SetTitle("#phi (degree)");
-  hTCPhiPos->GetYaxis()->SetTitle("Entries");
-  hTCPhiPos->SetLineColor(kBlue);
-  TH1D *hTCPhiNeg = new TH1D("hTCPhiNeg","#phi distribution of TCs for #eta<0",400,-200.,200.);
-  hTCPhiNeg->GetXaxis()->SetTitle("#phi (degree)");
-  hTCPhiNeg->GetYaxis()->SetTitle("Entries");
-  hTCPhiNeg->SetLineColor(kBlue);
-  TH2D *h2TCEtaPhi = new TH2D("h2TCEtaPhi","#eta-#phi distribution of TCs", 400,-200.,200., 1000,-4.15,4.15);
-  h2TCEtaPhi->GetXaxis()->SetTitle("#phi");
-  h2TCEtaPhi->GetYaxis()->SetTitle("#eta");
   //////////////////////////////////////////////////////////
   TH1D *tcRoz = new TH1D("tcRoz",Form("Roz distribution of TCs w.r.t gen"),1000,0.0,1.0);
-  TH1D *tcRozCore = new TH1D("tcRozCore",Form("Roz distribution of TCs w.r.t gen"),1000,0.0,1.0);
-  TH1D *htcMipPt = new TH1D("htcMipPt","tc mipPt",2500,0.0,2500.0);
-  TH1D *tcRozDet[7];
-  for(int idet=0;idet<7;idet++){
-    tcRozDet[idet] = new TH1D(Form("tcRozDet_%d",idet),Form("tcRozDet_%d",idet), 1000,0.0,1.0);
-  }
-  TH1D *tcRozPtwt = new TH1D("tcRozPtwt",Form("Roz distribution of TCs w.r.t gen"),1000,0.0,1.0);
   TH1D *clusRoz = new TH1D("clusRoz",Form("Roz distribution of Clusters w.r.t gen"),1000,0.0,1.0);
-  TH1D *clusRozPtwt = new TH1D("clusRozPtwt",Form("Roz distribution of Clusters w.r.t gen"),1000,0.0,1.0);
   double tcRoZCut = 0.15;
   ///////////////////////////////////////////////////////////
   TH1D *hGenvsJetPtDiff = new TH1D("hGenvsJetPtDiff",Form("(p_{T}^{gen}-p_{T}^{genJet}) GeV"),200,-100.0,100.0);
-  TProfile *pGenvsJetPtDiff = new TProfile("pGenvsJetPtDiff","p_{T}^{gen}/p_{T}^{genJet}", 100, 0., 500., 0., 10.);
+  TProfile *pGenvsJetPtDiff = new TProfile("pGenvsJetPtDiff","pGenvsJetPtDiff", 100, 0., 500., -10., 10.);
   TH1D *hGenvsJetDR = new TH1D("hGenvsJetDR",Form("#DeltaR"),800,-0.1,0.7);
-  TProfile *pPtDiffvsDR = new TProfile("pPtDiffvsDR",Form("(p_{T}^{gen}/p_{T}^{genJet}) vs #DeltaR"),800,-0.1,0.7,0.,10.);
-  TH2D *h2PtDiffvsDR = new TH2D("h2PtDiffvsDR",Form("(p_{T}^{gen}/p_{T}^{genJet}) vs #DeltaR"),800,-0.1,0.7,100, 0.,10.);
-
-  ////////////////////////////////////////////////////////////
   TH1D *hNJets = new TH1D("hNJets",Form("Nof VBF Jets"), 10,-0.5,9.5);
   TH1D *hNJetsF1 = new TH1D("hNJetsF1",Form("Nof VBF Jets (filtered for atmost 2)"), 10,-0.5,9.5);
-  ////////////////////////////////////////////////////////////
-
-  ///////////// Nof Clusters ////////////////////////////////
-  TH1D *hNClus1GeV = new TH1D("hNClus1GeV","Nof cluster with p_{T}>1 GeV per side", 50,-0.5,49.5);
-  TH1D *hNClus3GeV = new TH1D("hNClus3GeV","Nof cluster with p_{T}>3 GeV per side", 50,-0.5,49.5);
-  TH1D *hNClus5GeV = new TH1D("hNClus5GeV","Nof cluster with p_{T}>5 GeV per side ", 50,-0.5,49.5);
-  TH1D *hNClus10GeV = new TH1D("hNClus10GeV","Nof cluster with p_{T}>10 GeV per side", 50,-0.5,49.5);
-  //////////////////////////////////////////////////////////
   
   ///////////// TC Correction ///////////////////////////////////
   TF1* f2 = new TF1("func","[0] - [1]/pow(x,1/3) - [2]/pow(x,4/3) + [3]/pow(x,2) - [4]*x ",3.,3.5e4);
@@ -748,7 +665,7 @@ int main(int argc, char** argv)
   // fhigh->SetParameters(1.389, 7.78113e-05, 2,137);
   // TF1 *fhigh1 = new TF1("fhigh1","[0] * log (x-[1])",0.2e3,12.e3);
   // fhigh1->SetParameters(0.166198, -11.3154);
-  
+
   // TF1* f2 = new TF1(Form("func"),"[0] - [1]/pow(x,1/3) - [2]/pow(x,4/3) + [3]/pow(x,2) - [4]*log (x-[5]) + [6]*x - [7]/x",0.1,7.5e4);
   // f2->SetParameters(0.816674, 1.17338, -136.175, 0.498167, -0.0698533, -1362.15, 9.2412e-06, 136.382); //PU200 VBF CB with userange 0 to 0.8  
   // TF1 *fhigh = new TF1("fhigh","[0] * log (x-[1]) + [2] + [3]*x",0.5e3,12.e3);
@@ -757,9 +674,10 @@ int main(int argc, char** argv)
   ////////////////Cluster correction ///////////////////////////
   TF1* fClusCorr = new TF1("fClusCorr","[0]+[1]/(x+[2])",7,500);
   fClusCorr->SetParameters(1.22466, 22.8516, 0.401581); //PU0 VBF
-  
+
   TF1* fcl3d = new TF1("fcl3d","[0]+[1]/(x+[2])",7,500);
   fcl3d->SetParameters(1.17576, 19.7599, 0.908577); //PU0 VBF
+
   
   //////////////////////////////////////////////////////////
   uint64_t totalEntries = tr->GetEntries();
@@ -815,7 +733,7 @@ int main(int argc, char** argv)
       p.index = igen;
       genlist.push_back(p);
       
-      if(doPrint and igen<=10
+      if(doPrint and igen<=8
 	 // and
   	 // (gennom.find("h0")!=std::string::npos
   	 // or
@@ -832,7 +750,7 @@ int main(int argc, char** argv)
   		  << ", " << std::setw(10) << gen_eta->at(igen) << ", " << std::setw(10) << (TMath::RadToDeg()*gen_phi->at(igen)) << ", " << std::setw(10) << gen_energy->at(igen) << ") "
   		  <<" daughters: " << (gen_daughters->at(igen)).size()
   		  << ", " << daughterlist
-	  //<< ", taudlist.size(): " << taudlist.size() 
+		  << ", taudlist.size(): " << taudlist.size() 
   		  << std::defaultfloat
   		  << std::endl;
       }//print condition
@@ -868,18 +786,8 @@ int main(int argc, char** argv)
     
     std::vector<JetPart> jetlist;
     double delRTh = 0.25;
-    int nofHGCalJetsPos = 0, nofHGCalJetsNeg = 0;
-    int nofJetsLEPhiPos = 0, nofJetsLEPhiNeg = 0;
     for(int ijet=0; ijet<genjet_n; ijet++ ){
       if(fabs(genjet_eta->at(ijet))<=1.321 or fabs(genjet_eta->at(ijet))>=3.152) continue;
-      double phiDeg = TMath::RadToDeg()*genjet_phi->at(ijet);
-      if(genjet_eta->at(ijet)>0.){
-	nofHGCalJetsPos++;
-	if((phiDeg>=-95. and phiDeg<=-85.) or (phiDeg>=25. and phiDeg<=35.) or (phiDeg>=145. and phiDeg<=155.)) nofJetsLEPhiPos++;
-      }else{
-	nofHGCalJetsNeg++;
-	if((phiDeg>=-95. and phiDeg<=-85.) or (phiDeg>=25. and phiDeg<=35.) or (phiDeg>=145. and phiDeg<=155.)) nofJetsLEPhiNeg++;
-      }
       JetPart jet;
       // jet.name = "";
       // jet.index = ijet ;
@@ -893,19 +801,10 @@ int main(int argc, char** argv)
       for(int ipart=0; ipart<genlist.size(); ipart++){
 	int refpart = genlist.at(ipart).index;
 	double deltaR = TMath::Sqrt((gen_eta->at(refpart)-genjet_eta->at(ijet))*(gen_eta->at(refpart)-genjet_eta->at(ijet)) + (gen_phi->at(refpart)-genjet_phi->at(ijet))*(gen_phi->at(refpart)-genjet_phi->at(ijet)));
-	bool jetcondn = (deltaR<delRTh and deltaR<minDeltaR);
-	double deltaR_gen5 = 1., deltaR_gen6 = 1., deltaR_gen7 = 1.;
-	double ptratio = 1.;
-	if(sampleType.find("vbfHInv")!=std::string::npos) {
-	  deltaR_gen5 = TMath::Sqrt( (gen_eta->at(5)-genjet_eta->at(ijet))*(gen_eta->at(5)-genjet_eta->at(ijet)) + (gen_phi->at(5)-genjet_phi->at(ijet))*(gen_phi->at(5)-genjet_phi->at(ijet)) );
-	  deltaR_gen6 = TMath::Sqrt( (gen_eta->at(6)-genjet_eta->at(ijet))*(gen_eta->at(6)-genjet_eta->at(ijet)) + (gen_phi->at(6)-genjet_phi->at(ijet))*(gen_phi->at(6)-genjet_phi->at(ijet)) );
-	  deltaR_gen7 = TMath::Sqrt( (gen_eta->at(7)-genjet_eta->at(ijet))*(gen_eta->at(7)-genjet_eta->at(ijet)) + (gen_phi->at(7)-genjet_phi->at(ijet))*(gen_phi->at(7)-genjet_phi->at(ijet)) );
-	  ptratio = gen_pt->at(refpart)/genjet_pt->at(ijet);
-	  //jetcondn = (jetcondn and (deltaR_gen5<delRTh or deltaR_gen6<delRTh or deltaR_gen7<delRTh));
-	  jetcondn = (jetcondn and (deltaR_gen5<delRTh or deltaR_gen6<delRTh or deltaR_gen7<delRTh) and (ptratio>0.8 and ptratio<1.2));
-	}
-	if(jetcondn){
-	  //if(deltaR<delRTh and deltaR<minDeltaR and (deltaR_gen5<delRTh or deltaR_gen6<delRTh or deltaR_gen7<delRTh) ){
+	double deltaR_gen5 = TMath::Sqrt( (gen_eta->at(5)-genjet_eta->at(ijet))*(gen_eta->at(5)-genjet_eta->at(ijet)) + (gen_phi->at(5)-genjet_phi->at(ijet))*(gen_phi->at(5)-genjet_phi->at(ijet)) );
+	double deltaR_gen6 = TMath::Sqrt( (gen_eta->at(6)-genjet_eta->at(ijet))*(gen_eta->at(6)-genjet_eta->at(ijet)) + (gen_phi->at(6)-genjet_phi->at(ijet))*(gen_phi->at(6)-genjet_phi->at(ijet)) );
+	double deltaR_gen7 = TMath::Sqrt( (gen_eta->at(7)-genjet_eta->at(ijet))*(gen_eta->at(7)-genjet_eta->at(ijet)) + (gen_phi->at(7)-genjet_phi->at(ijet))*(gen_phi->at(7)-genjet_phi->at(ijet)) );	
+	if(deltaR<delRTh and deltaR<minDeltaR and (deltaR_gen5<delRTh or deltaR_gen6<delRTh or deltaR_gen7<delRTh)){
 	  //if(deltaR<0.4 and deltaR<minDeltaR){
 	  minDeltaR = deltaR;
 	  minDeltaRindex = ipart;
@@ -924,219 +823,151 @@ int main(int argc, char** argv)
 	  jet.index = ijet ;
 	  jetlist.push_back(jet);
 	  
-	  if(sampleType.find("vbfHInv")!=std::string::npos) {
-	    int mindr = TMath::LocMin(3,minDR);
-	    int minId = mindr + 5; //index of non-Higgs particles of the hardest subprocess in VBF sample
-	    hGenvsJetPtDiff->Fill(gen_pt->at(minId)-genjet_pt->at(ijet));	  
-	    pGenvsJetPtDiff->Fill( gen_pt->at(minId),  gen_pt->at(minId)/genjet_pt->at(ijet) );
-	    pPtDiffvsDR->Fill( minDR[mindr],  gen_pt->at(minId)/genjet_pt->at(ijet) );
-	    h2PtDiffvsDR->Fill( minDR[mindr],  gen_pt->at(minId)/genjet_pt->at(ijet) );
-	    hGenvsJetDR->Fill( minDR[mindr]  );
-	  }else{
-	    hGenvsJetPtDiff->Fill(gen_pt->at(ipart)-genjet_pt->at(ijet));	  
-	    pGenvsJetPtDiff->Fill( gen_pt->at(ipart),  gen_pt->at(ipart)/genjet_pt->at(ijet) );
-	    pPtDiffvsDR->Fill( minDeltaR,  gen_pt->at(ipart)/genjet_pt->at(ijet) );
-	    h2PtDiffvsDR->Fill( minDeltaR,  gen_pt->at(ipart)/genjet_pt->at(ijet) );
-	    hGenvsJetDR->Fill( minDeltaR  );
-	  }
-	  if(doPrint){
-	    if(sampleType.find("vbfHInv")!=std::string::npos)
-	      std::cout << "ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
-			<<", Name: " << std::setw(10) << p.name
-			<< ", deltaR : " << std::setprecision(3) << std::setw(5) << minDeltaR
-			<< ", deltaR5 : " << std::setprecision(3) << std::setw(5) << minDR[0]
-			<< ", deltaR6 : " << std::setprecision(3) << std::setw(5) << minDR[1]
-			<< ", deltaR7 : " << std::setprecision(3) << std::setw(5) << minDR[2]
-			<<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
-			<< ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
-			<< ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
-			<< ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
-			<< std::defaultfloat
-			<< std::endl;
-	    else
-	      std::cout << "ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
-			<<", Name: " << std::setw(10) << p.name
-			<< ", deltaR : " << std::setprecision(3) << std::setw(5) << minDeltaR
-			<<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
-			<< ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
-			<< ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
-			<< ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
-			<< std::defaultfloat
-			<< std::endl;
-
-	  }//print condn
+	  int mindr = TMath::LocMin(3,minDR);
+	  int minId = mindr + 5; //index of non-Higgs particles of the hardest subprocess in VBF sample
+	  hGenvsJetPtDiff->Fill(gen_pt->at(minId)-genjet_pt->at(ijet));	  
+	  pGenvsJetPtDiff->Fill( gen_pt->at(minId),  gen_pt->at(minId)/genjet_pt->at(ijet) );
+	  hGenvsJetDR->Fill( minDR[mindr]  );
+	  if(doPrint)
+	    std::cout << "ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
+		      <<", Name: " << std::setw(10) << p.name << ", deltaR : " << std::setprecision(3) << std::setw(5) << minDeltaR
+		      << ", deltaR5 : " << std::setprecision(3) << std::setw(5) << minDR[0]
+		      << ", deltaR6 : " << std::setprecision(3) << std::setw(5) << minDR[1]
+		      << ", deltaR7 : " << std::setprecision(3) << std::setw(5) << minDR[2]
+		      <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
+		      << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
+		      << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
+		      << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
+		      // <<" jet:(roz,xoz,yoz) : (" << std::fixed << std::setprecision(2) << std::setw(5) << roz
+		      // <<", " << std::fixed << std::setprecision(2) << std::setw(5) << xoz
+		      // <<", " << std::fixed << std::setprecision(2) << std::setw(5) << yoz
+		      // << ") "
+		      << std::defaultfloat
+		      << std::endl;
 	}//minDelta condition
       }//part loop
     }//jet loop
     /////////////////////////////////////////////////////////////////////////////////////////
     hNJets->Fill(float(jetlist.size()));
 
-    ////////////////////////Conditions for two jets ////////////////////////////////////
-    bool isOneperSide = true ;
+    /////////////////////////// Conditions for two jets ////////////////////////////////////
     if(jetlist.size()==2){
-      isOneperSide = ((genjet_eta->at(jetlist.at(0).index) * genjet_eta->at(jetlist.at(1).index))<0.)?true:false;
-      // double *allJets = new double[jetlist.size()];
-      // int *sortedIdx = new int[jetlist.size()];
-      // for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-      // 	int ijet = jetlist.at(ipjet).index ;
-      // 	allJets[ipjet] = genjet_pt->at(ijet);
-      // }
-      // TMath::Sort(int(jetlist.size()), allJets, sortedIdx);
-      // int HptIdx = jetlist.at(sortedIdx[0]).index ;
-      // int MptIdx = jetlist.at(sortedIdx[1]).index ;
-      // double eta12 = genjet_eta->at(HptIdx) * genjet_eta->at(MptIdx) ;
-      // std::vector<JetPart> tJetlist;
-      // for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-      // 	if(eta12<0.){
-      // 	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[1]){
-      // 	    JetPart jet;
-      // 	    jet.name = jetlist.at(ipjet).name ;
-      // 	    jet.index = jetlist.at(ipjet).index ;
-      // 	    tJetlist.push_back(jet);
-      // 	  }
-      // 	}else{
-      // 	  if(ipjet==sortedIdx[0]){
-      // 	    JetPart jet;
-      // 	    jet.name = jetlist.at(ipjet).name ;
-      // 	    jet.index = jetlist.at(ipjet).index ;
-      // 	    tJetlist.push_back(jet);
-      // 	  }	  
-      // 	}
-      // }
-      // jetlist.clear();
-      // for(int ipjet=0; ipjet < tJetlist.size() ; ipjet++ ){
-      // 	int ijet = tJetlist.at(ipjet).index ;
-      // 	JetPart jet;
-      // 	jet.name = tJetlist.at(ipjet).name ;
-      // 	jet.index = tJetlist.at(ipjet).index ;
-      // 	jetlist.push_back(jet);
-      // 	if(doPrint)
-      // 	  std::cout << "filtered-ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
-      // 		    <<", Name: " << std::setw(10) << jet.name
-      // 		    <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
-      // 		    << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
-      // 		    << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
-      // 		    << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
-      // 		    << std::defaultfloat
-      // 		    << std::endl;
-      // }      
-      // tJetlist.clear();      
-      // delete allJets;
-      // delete sortedIdx;
-    }
-    ////////////////////////////////////////////////////////////////////////////////////////
-    
-    // ///////////////////////// Conditions for three jets ////////////////////////////////////
-    // if(jetlist.size()>2){
-    //   double *allJets = new double[jetlist.size()];
-    //   int *sortedIdx = new int[jetlist.size()];
-    //   for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-    // 	int ijet = jetlist.at(ipjet).index ;
-    // 	allJets[ipjet] = genjet_pt->at(ijet);
-    //   }
-    //   TMath::Sort(int(jetlist.size()), allJets, sortedIdx);
-    //   int HptIdx = jetlist.at(sortedIdx[0]).index ;
-    //   int MptIdx = jetlist.at(sortedIdx[1]).index ;
-    //   int SptIdx = jetlist.at(sortedIdx[2]).index ;
-    //   double eta12 = genjet_eta->at(HptIdx) * genjet_eta->at(MptIdx) ;
-    //   double eta13 = genjet_eta->at(HptIdx) * genjet_eta->at(SptIdx) ;
-    //   std::vector<JetPart> tJetlist;
-    //   for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-    // 	if(eta12<0.){
-    // 	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[1]){
-    // 	    JetPart jet;
-    // 	    jet.name = jetlist.at(ipjet).name ;
-    // 	    jet.index = jetlist.at(ipjet).index ;
-    // 	    tJetlist.push_back(jet);
-    // 	  }
-    // 	}else if(eta13<0.){
-    // 	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[2]){
-    // 	    JetPart jet;
-    // 	    jet.name = jetlist.at(ipjet).name ;
-    // 	    jet.index = jetlist.at(ipjet).index ;
-    // 	    tJetlist.push_back(jet);
-    // 	  }
-    // 	}else{
-    // 	  if(ipjet==sortedIdx[0]){
-    // 	    JetPart jet;
-    // 	    jet.name = jetlist.at(ipjet).name ;
-    // 	    jet.index = jetlist.at(ipjet).index ;
-    // 	    tJetlist.push_back(jet);
-    // 	  }	  
-    // 	}
-    //   }
-    //   jetlist.clear();
-    //   for(int ipjet=0; ipjet < tJetlist.size() ; ipjet++ ){
-    // 	int ijet = tJetlist.at(ipjet).index ;
-    // 	JetPart jet;
-    // 	jet.name = tJetlist.at(ipjet).name ;
-    // 	jet.index = tJetlist.at(ipjet).index ;
-    // 	jetlist.push_back(jet);
-    // 	if(doPrint)
-    // 	  std::cout << "filtered-ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
-    // 		    <<", Name: " << std::setw(10) << jet.name
-    // 		    <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
-    // 		    << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
-    // 		    << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
-    // 		    << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
-    // 		    << std::defaultfloat
-    // 		    << std::endl;
-    //   }      
-    //   tJetlist.clear();      
-    //   delete allJets;
-    //   delete sortedIdx;
-    // }
-    // ////////////////////////////////////////////////////////////////////////////////////////
-
-    ////////////////////////////////////////////////////////////////////////////////////////
-    double xoz_posEta = -10., yoz_posEta = -10.;
-    double xoz_negEta = -10., yoz_negEta = -10.;
-    double genjetpt_posEta =  -1., genjetpt_negEta = -1.;
-    int njets = 0, njetsp = 0, njetsm = 0;
-    for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-      int ijet = jetlist.at(ipjet).index ;
-      if(fabs(genjet_eta->at(ijet))<=1.321 or fabs(genjet_eta->at(ijet))>=3.152) continue;
-      double roz = 1/sinh(genjet_eta->at(ijet)) ; //tan(2*atan(exp(-1.0*genjet_eta->at(ijet))));
-      double xoz = roz*cos(genjet_phi->at(ijet));
-      double yoz = roz*sin(genjet_phi->at(ijet));
-      if(genjet_eta->at(ijet)>0. and nofHGCalJetsPos==1 ){ //and nofJetsLEPhiPos==0
-	xoz_posEta = xoz; yoz_posEta = yoz;
-	genjetpt_posEta = genjet_pt->at(ijet);
-	njets++;
-	njetsp++;
-      }else if(genjet_eta->at(ijet)<0. and nofHGCalJetsNeg==1 ){ //and nofJetsLEPhiNeg==0
-	xoz_negEta = xoz; yoz_negEta = yoz;
-	genjetpt_negEta = genjet_pt->at(ijet);
-	njets++;
-	njetsm++;
+      double *allJets = new double[jetlist.size()];
+      int *sortedIdx = new int[jetlist.size()];
+      for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
+	int ijet = jetlist.at(ipjet).index ;
+	allJets[ipjet] = genjet_pt->at(ijet);
       }
-      if(doPrint)
-	std::cout << "ijet1-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
-		  <<", Name: " << std::setw(10) << jetlist.at(ipjet).name
-		  <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
-		  << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
-		  << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
-		  << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
-		  <<" jet:(roz,xoz,yoz) : (" << std::fixed << std::setprecision(2) << std::setw(5) << roz
-		  <<", " << std::fixed << std::setprecision(2) << std::setw(5) << xoz
-		  <<", " << std::fixed << std::setprecision(2) << std::setw(5) << yoz
-		  << ") "
-		  << std::defaultfloat
-		  << std::endl;
-      
+      TMath::Sort(int(jetlist.size()), allJets, sortedIdx);
+      int HptIdx = jetlist.at(sortedIdx[0]).index ;
+      int MptIdx = jetlist.at(sortedIdx[1]).index ;
+      double eta12 = genjet_eta->at(HptIdx) * genjet_eta->at(MptIdx) ;
+      std::vector<JetPart> tJetlist;
+      for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
+	if(eta12<0.){
+	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[1]){
+	    JetPart jet;
+	    jet.name = jetlist.at(ipjet).name ;
+	    jet.index = jetlist.at(ipjet).index ;
+	    tJetlist.push_back(jet);
+	  }
+	}else{
+	  if(ipjet==sortedIdx[0]){
+	    JetPart jet;
+	    jet.name = jetlist.at(ipjet).name ;
+	    jet.index = jetlist.at(ipjet).index ;
+	    tJetlist.push_back(jet);
+	  }	  
+	}
+      }
+      jetlist.clear();
+      for(int ipjet=0; ipjet < tJetlist.size() ; ipjet++ ){
+	int ijet = tJetlist.at(ipjet).index ;
+	JetPart jet;
+	jet.name = tJetlist.at(ipjet).name ;
+	jet.index = tJetlist.at(ipjet).index ;
+	jetlist.push_back(jet);
+	if(doPrint)
+	  std::cout << "filtered-ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
+		    <<", Name: " << std::setw(10) << jet.name
+		    <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
+		    << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
+		    << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
+		    << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
+		    << std::defaultfloat
+		    << std::endl;
+      }      
+      tJetlist.clear();      
+      delete allJets;
+      delete sortedIdx;
     }
     ////////////////////////////////////////////////////////////////////////////////////////
     
-    if(njets>2){
-      std::cout << "Issue: Event: " << ievent << ", njetsp: " << njetsp << ", njetsm: " << njetsm << std::endl;
+    ///////////////////////// Conditions for three jets ////////////////////////////////////
+    if(jetlist.size()>2){
+      double *allJets = new double[jetlist.size()];
+      int *sortedIdx = new int[jetlist.size()];
+      for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
+	int ijet = jetlist.at(ipjet).index ;
+	allJets[ipjet] = genjet_pt->at(ijet);
+      }
+      TMath::Sort(int(jetlist.size()), allJets, sortedIdx);
+      int HptIdx = jetlist.at(sortedIdx[0]).index ;
+      int MptIdx = jetlist.at(sortedIdx[1]).index ;
+      int SptIdx = jetlist.at(sortedIdx[2]).index ;
+      double eta12 = genjet_eta->at(HptIdx) * genjet_eta->at(MptIdx) ;
+      double eta13 = genjet_eta->at(HptIdx) * genjet_eta->at(SptIdx) ;
+      std::vector<JetPart> tJetlist;
+      for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
+	if(eta12<0.){
+	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[1]){
+	    JetPart jet;
+	    jet.name = jetlist.at(ipjet).name ;
+	    jet.index = jetlist.at(ipjet).index ;
+	    tJetlist.push_back(jet);
+	  }
+	}else if(eta13<0.){
+	  if(ipjet==sortedIdx[0] or ipjet==sortedIdx[2]){
+	    JetPart jet;
+	    jet.name = jetlist.at(ipjet).name ;
+	    jet.index = jetlist.at(ipjet).index ;
+	    tJetlist.push_back(jet);
+	  }
+	}else{
+	  if(ipjet==sortedIdx[0]){
+	    JetPart jet;
+	    jet.name = jetlist.at(ipjet).name ;
+	    jet.index = jetlist.at(ipjet).index ;
+	    tJetlist.push_back(jet);
+	  }	  
+	}
+      }
+      jetlist.clear();
+      for(int ipjet=0; ipjet < tJetlist.size() ; ipjet++ ){
+	int ijet = tJetlist.at(ipjet).index ;
+	JetPart jet;
+	jet.name = tJetlist.at(ipjet).name ;
+	jet.index = tJetlist.at(ipjet).index ;
+	jetlist.push_back(jet);
+	if(doPrint)
+	  std::cout << "filtered-ijet-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
+		    <<", Name: " << std::setw(10) << jet.name
+		    <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
+		    << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
+		    << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
+		    << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
+		    << std::defaultfloat
+		    << std::endl;
+      }      
+      tJetlist.clear();      
+      delete allJets;
+      delete sortedIdx;
     }
+    ////////////////////////////////////////////////////////////////////////////////////////
     
-    /////////////////////////// Conditions for 0 jets and others ////////////////////////////////////
-
-    if(njets==0 or (jetlist.size()==2 and !isOneperSide) or njets>2){
-      //if(jetlist.size()==0 or !isOneperSide or nofHGCalJetsPos>1 or nofHGCalJetsNeg>1){
-      //if(jetlist.size()==0 or jetlist.size()>2 or !isOneperSide){
-      //if(jetlist.size()==0){
+    
+    /////////////////////////// Conditions for 0 jets ////////////////////////////////////
+    if(jetlist.size()==0){
       taudlist.clear();
       taugdlist.clear();
       genlist.clear();
@@ -1167,7 +998,6 @@ int main(int argc, char** argv)
       genjet_phi->clear();
       genjet_energy->clear();
       tc_pt->clear();
-      tc_mipPt->clear();
       tc_eta->clear();
       tc_phi->clear();
       tc_layer->clear();
@@ -1186,8 +1016,41 @@ int main(int argc, char** argv)
     }
     ////////////////////////////////////////////////////////////////////////////////////////
     
-    hNJetsF1->Fill(float(njets));
-    
+    hNJetsF1->Fill(float(jetlist.size()));
+
+    ////////////////////////////////////////////////////////////////////////////////////////
+    double xoz_posEta = -10., yoz_posEta = -10.;
+    double xoz_negEta = -10., yoz_negEta = -10.;
+    double genjetpt_posEta =  -1., genjetpt_negEta = -1.;
+    for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
+      int ijet = jetlist.at(ipjet).index ;
+      double roz = 1/sinh(genjet_eta->at(ijet)) ; //tan(2*atan(exp(-1.0*genjet_eta->at(ijet))));
+      double xoz = roz*cos(genjet_phi->at(ijet));
+      double yoz = roz*sin(genjet_phi->at(ijet));
+      if(genjet_eta->at(ijet)>0){
+	xoz_posEta = xoz; yoz_posEta = yoz;
+	genjetpt_posEta = genjet_pt->at(ijet);
+      }else{
+	xoz_negEta = xoz; yoz_negEta = yoz;
+	genjetpt_negEta = genjet_pt->at(ijet);
+      }
+      if(doPrint)
+	std::cout << "ijet1-ievent: " << std::setprecision(default_precision) << std::setw(4) <<ievent << ", ijet: " << std::setw(4) << ijet
+		  <<", Name: " << std::setw(10) << jetlist.at(ipjet).name
+		  <<" jet:(pt,eta,phi,energy) : (" << std::fixed << std::setprecision(2) << std::setw(8) << genjet_pt->at(ijet)
+		  << ", " << std::setw(8) << genjet_eta->at(ijet) << ", " //<< std::setw(5) << genjet_exeta->at(ijet)
+		  << ", " << std::setw(8) << (TMath::RadToDeg()*genjet_phi->at(ijet)) //<< ", " << std::setw(5) << (TMath::RadToDeg()*genjet_exphi->at(ijet))
+		  << ", " << std::setw(8) << genjet_energy->at(ijet) << ") "
+		  <<" jet:(roz,xoz,yoz) : (" << std::fixed << std::setprecision(2) << std::setw(5) << roz
+		  <<", " << std::fixed << std::setprecision(2) << std::setw(5) << xoz
+		  <<", " << std::fixed << std::setprecision(2) << std::setw(5) << yoz
+		  << ") "
+		  << std::defaultfloat
+		  << std::endl;
+      
+    }
+
+    ////////////////////////////////////////////////////////////////////////////////////////
     
     // std::vector<TCOrder> rearrgdtcs;
     // for(unsigned itc=0;itc<tc_pt->size();itc++){
@@ -1205,12 +1068,7 @@ int main(int argc, char** argv)
     /////////////////////////////////////////////////////////////////////////////////////////
     double totTCpt_posEta_iter1 = 0.,totTCpt_negEta_iter1 = 0.;
     double totTCpt_posEta_CEE_iter1 = 0.,totTCpt_negEta_CEE_iter1 = 0.;
-    double totTCpt_posEta_core_iter1 = 0.,totTCpt_negEta_core_iter1 = 0.;
-    double totTCpt_posEta_core_rozcut0p15_iter1 = 0.,totTCpt_negEta_core_rozcut0p15_iter1 = 0.;
-    double totTCpt_posEta_core_mpptcut2_iter1 = 0.,totTCpt_negEta_core_mpptcut2_iter1 = 0.;
-    double totTCpt_posEta_CEE_core_iter1 = 0.,totTCpt_negEta_CEE_core_iter1 = 0.;
     for(unsigned itc=0;itc<tc_pt->size();itc++){
-      
       double minRoz = 1., clstPt = 0, clstEta = 0;
       for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
 	int ijet = jetlist.at(ipjet).index ;
@@ -1226,8 +1084,7 @@ int main(int argc, char** argv)
       }
       double tcptMeV = tc_pt->at(itc) * 1.e3;
       double tcpt = (tcptMeV<3.)?3:tcptMeV; //VBF PU0      
-      double ptcorr = (tc_layer->at(itc)<=26)?1.07:1.14;//1.13 for Hadronic part //f2->Eval(tcpt);
-      //double ptcorr = f2->Eval(tcpt);
+      double ptcorr = f2->Eval(tcpt);
       
       // double tcpt = (tcptMeV<0.1)?0.1:tcptMeV; //VBF PU200
       // double ptcorr_high = fhigh->Eval(tcpt);
@@ -1235,54 +1092,24 @@ int main(int argc, char** argv)
       // if(tcpt>400.) ptcorr = ptcorr* ptcorr_high;
       // // if(tcpt>200.) ptcorr = ptcorr* ptcorr_high1;
       
-      tcRoz->Fill(minRoz);
-      htcMipPt->Fill(tc_mipPt->at(itc));
-      tcRozPtwt->Fill(minRoz, tc_pt->at(itc));
-
-      //if(fabs(clstEta)>etaMinCore and fabs(clstEta)<etaMaxCore){ 
-	if(tc_eta->at(itc)<0 and genjetpt_negEta>-1.){
+      tcRoz->Fill(minRoz, tc_pt->at(itc));
+      if(fabs(clstEta)>1.6 and fabs(clstEta)<3.0 and minRoz<tcRoZCut){ //only add for the plateau region	
+	if(tc_eta->at(itc)<0 and genjetpt_negEta>=0.){
 	  totTCpt_negEta_iter1 += ptcorr*tc_pt->at(itc);
-	  if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE_iter1 += ptcorr*tc_pt->at(itc);    
-	}else if(tc_eta->at(itc)>0. and genjetpt_posEta>-1.){
+	  if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE_iter1 += tc_pt->at(itc);
+	}else if(tc_eta->at(itc)>0. and genjetpt_posEta>=0.){
+	  //std::cout <<", clstEta: " << std::fixed << std::setprecision(2) << std::setw(8) << clstEta << std::defaultfloat << std::endl;
 	  totTCpt_posEta_iter1 += ptcorr*tc_pt->at(itc);
-	  if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE_iter1 += ptcorr*tc_pt->at(itc);
-	}
-	//}
-      
-      if(fabs(clstEta)>etaMinCore and fabs(clstEta)<etaMaxCore){ //only add for the plateau region	
-	if(tc_eta->at(itc)<0 and genjetpt_negEta>-1.){
-	  totTCpt_negEta_core_iter1 += ptcorr*tc_pt->at(itc);
-	  if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE_core_iter1 += ptcorr*tc_pt->at(itc);
-	}else if(tc_eta->at(itc)>0. and genjetpt_posEta>-1.){
-	  totTCpt_posEta_core_iter1 += ptcorr*tc_pt->at(itc);
-	  if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE_core_iter1 += ptcorr*tc_pt->at(itc);
+	  if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE_iter1 += tc_pt->at(itc);
 	}
       }
-      
-      if(fabs(clstEta)>etaMinCore and fabs(clstEta)<etaMaxCore and minRoz<tcRoZCut){ //only add for the plateau region
-	tcRozCore->Fill(minRoz);
-	if(tc_eta->at(itc)<0 and genjetpt_negEta>-1.){
-	  totTCpt_negEta_core_rozcut0p15_iter1 += ptcorr*tc_pt->at(itc);
-	}else if(tc_eta->at(itc)>0. and genjetpt_posEta>-1.){
-	  totTCpt_posEta_core_rozcut0p15_iter1 += ptcorr*tc_pt->at(itc);
-	}
-      }
-      
-      if(fabs(clstEta)>etaMinCore and fabs(clstEta)<etaMaxCore and tc_mipPt->at(itc)>2.){ //only add for the plateau region	
-	if(tc_eta->at(itc)<0 and genjetpt_negEta>-1.){
-	  totTCpt_negEta_core_mpptcut2_iter1 += ptcorr*tc_pt->at(itc);	  
-	}else if(tc_eta->at(itc)>0. and genjetpt_posEta>-1.){
-	  totTCpt_posEta_core_mpptcut2_iter1 += ptcorr*tc_pt->at(itc);
-	}
-      }
-      
-      
     }    
     /////////////////////////////////////////////////////////////////////////////////////////
+    float tot_tc_pt = 0.0, tot_tc_e = 0.0;
     double totTCpt_posEta = 0.,totTCpt_negEta = 0.;
     double totTCpt_posEta_CEE = 0.,totTCpt_negEta_CEE = 0.;
     double totTCpt_posEta_layer1 = 0.,totTCpt_negEta_layer1 = 0.;
-    uint32_t nofPosEtaTCs = 0, nofNegEtaTCs = 0;
+    uint32_t nofPosEtaTCs = 0, nofNegEtaTCs = 0, nofTCsTrigL2 = 0, nofTCsTrigL9 = 0;
     double tcXoZposEta = 0.,tcYoZposEta = 0., tcXoZnegEta = 0.,tcYoZnegEta = 0.;
     double maxroz_posEta = -1.0, maxroz_negEta = -1.0;
     double tcPhi_posEta = 0.,tcPhi_negEta = 0.;
@@ -1363,12 +1190,9 @@ int main(int argc, char** argv)
 	  clstEta = genjet_eta->at(ijet);
 	}
       }
-      tcRozDet[detType-1]->Fill(minRoz);
-      
       double tcptMeV = tc_pt->at(itc) * 1.e3;
       double tcpt = (tcptMeV<3.)?3:tcptMeV;
-      double ptcorr = (tc_layer->at(itc)<=26)?1.07:1.14;//1.07:1.13f2->Eval(tcpt);
-      //double ptcorr = f2->Eval(tcpt);
+      double ptcorr = f2->Eval(tcpt);
       tcpt = ptcorr*tc_pt->at(itc);
       tcptMeV = tcpt*1.e3;
       
@@ -1399,32 +1223,26 @@ int main(int argc, char** argv)
 	  tcf0.setEnergyGeV(scale * tc_pt->at(itc));
 	  tcf0.setLayer(tc_layer->at(itc));
 	  //if(doPrint) tcf0.print();
-	  //vTcw[isect+addisect].push_back(tcf0);
-	  if((tc_eta->at(itc)<0. and genjetpt_negEta>-1.) or (tc_eta->at(itc)>0. and genjetpt_posEta>-1.))
-	    vTcw[isect+addisect].push_back(tcf0);
+	  vTcw[isect+addisect].push_back(tcf0);
 	}
       }
       
       // bool hgcalEta = (fabs(tc_eta->at(itc))>1.321 and fabs(tc_eta->at(itc))<3.152) ? true : false;
-      //bool hgcalInnerEta = (fabs(tc_eta->at(itc))>etaMinCore and fabs(tc_eta->at(itc))<etaMaxCore) ? true : false;
-      bool hgcalInnerEta = (fabs(clstEta)>etaMinCore and fabs(clstEta)<etaMaxCore) ? true : false;
+      //bool hgcalInnerEta = (fabs(tc_eta->at(itc))>1.6 and fabs(tc_eta->at(itc))<3.0) ? true : false;
+      bool hgcalInnerEta = (fabs(clstEta)>1.6 and fabs(clstEta)<3.0) ? true : false;
       int binEta = 1, binPt = 1, binPtLog = 1;
       GetHistoBin(hJetEtaTCPtBin, tc_eta->at(itc), tcpt, binEta, binPt);
       GetHistoBin(hJetEtaTCPtBinLog, tc_eta->at(itc), tcptMeV, binEta, binPtLog);
       
-      hTCPt->Fill(tc_pt->at(itc));
-      hTCEta->Fill(tc_eta->at(itc));
-      h2TCEtaPhi->Fill(TMath::RadToDeg()*tc_phi->at(itc), tc_eta->at(itc));
-      if(tc_eta->at(itc)<0. and genjetpt_negEta>-1.){
-	hTCPhiNeg->Fill(TMath::RadToDeg()*tc_phi->at(itc));
+      if(tc_eta->at(itc)<0. and genjetpt_negEta>=0.){
 	//if(hgcalInnerEta){
 	// totTCpt_negEta += tc_pt->at(itc);
 	// if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE += tc_pt->at(itc);
 	//}
-	//if(minRoz<tcRoZCut){
-	totTCpt_negEta += tcpt;
-	if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE += tcpt;
-	//}
+	if(minRoz<tcRoZCut){
+	  totTCpt_negEta += tcpt;
+	  if(tc_layer->at(itc)<=26) totTCpt_negEta_CEE += tcpt;
+	}
 	tcXoZnegEta += tcpt*tc_x->at(itc)/tc_z->at(itc) ;
 	tcYoZnegEta += tcpt*tc_y->at(itc)/tc_z->at(itc) ;
 	tcPhi_negEta += tcpt*tc_phi->at(itc) ;
@@ -1447,41 +1265,20 @@ int main(int argc, char** argv)
 	  pTCPtvsGenPtByTCPtsum->Fill(tcptMeV, (clstPt/totTCpt_negEta_iter1));
 	  pTCPtvsGenPtByTCPtsumLogx->Fill(tcptMeV, (clstPt/totTCpt_negEta_iter1));
 	  pTCPtvsGenPtByTCPtsumDet[detType-1]->Fill(tcptMeV, (clstPt/totTCpt_negEta_iter1));
+	  hGenTCPtResovsTCPt1D[binPt-1]->Fill( (clstPt - totTCpt_negEta_iter1) );
+	  hGenTCPtResovsTCPt1DLog[binPtLog-1]->Fill( (clstPt - totTCpt_negEta_iter1) );
 	  hGenjetByTCPtSumVsTCPt1DLog[binPtLog-1]->Fill( clstPt/totTCpt_negEta_iter1 );
-     	  hGenjetByTCPtSumVsTCPt2DDetLog[detType-1][binPtLog-1]->Fill( clstPt/totTCpt_negEta_iter1 );
-	  hGenjetByTCPtSumVsTCPt2DLog[binEta-1][binPtLog-1]->Fill( clstPt/totTCpt_negEta_iter1 );
-	  //hGenjetByTCPtSumVsTCPt3DLog[detType-1][binEta-1][binPtLog-1]->Fill( clstPt/totTCpt_negEta_iter1 );
+     	  hGenTCPtResovsTCPt2D[detType-1][binPt-1]->Fill( (clstPt - totTCpt_negEta_iter1) );
 	}
-	h2GenjetTCPtsumvsGenPt->Fill(clstPt, clstPt/totTCpt_negEta_iter1); 
-	h2GenjetTCPtsumvsGenEta->Fill(fabs(clstEta), clstPt/totTCpt_negEta_iter1);
-	pGenjetTCPtsumvsGenPt->Fill(clstPt, clstPt/totTCpt_negEta_iter1); 
-	pGenjetTCPtsumvsGenEta->Fill(fabs(clstEta), clstPt/totTCpt_negEta_iter1);
-	if(hgcalInnerEta){
-	  h2GenjetTCPtsumvsGenPt_CoreEta->Fill(clstPt, clstPt/totTCpt_negEta_CEE_core_iter1);
-	  pGenjetTCPtsumvsGenPt_CoreEta->Fill(clstPt, clstPt/totTCpt_negEta_CEE_core_iter1);
-	}
-	if(tc_layer->at(itc)<=26){
-	  h2GenjetTCPtsumvsGenPt_CEE->Fill(clstPt, clstPt/totTCpt_negEta_CEE_iter1);
-	  h2GenjetTCPtsumvsGenEta_CEE->Fill(fabs(clstEta), clstPt/totTCpt_negEta_CEE_iter1);
-	  pGenjetTCPtsumvsGenPt_CEE->Fill(clstPt, clstPt/totTCpt_negEta_CEE_iter1);
-	  pGenjetTCPtsumvsGenEta_CEE->Fill(fabs(clstEta), clstPt/totTCpt_negEta_CEE_iter1);
-	  if(hgcalInnerEta){
-	    h2GenjetTCPtsumvsGenPt_CEE_CoreEta->Fill(clstPt, clstPt/totTCpt_negEta_CEE_core_iter1);
-	    pGenjetTCPtsumvsGenPt_CEE_CoreEta->Fill(clstPt, clstPt/totTCpt_negEta_CEE_core_iter1);
-	    h2GenjetTCPtsumvsTCPtsum_CEE_CoreEta->Fill(totTCpt_negEta_CEE_core_iter1, clstPt/totTCpt_negEta_CEE_core_iter1);
-	    pGenjetTCPtsumvsTCPtsum_CEE_CoreEta->Fill(totTCpt_negEta_CEE_core_iter1, clstPt/totTCpt_negEta_CEE_core_iter1);	    
-	  }
-	}
-      }else if(tc_eta->at(itc)>0. and genjetpt_posEta>-1.){
+      }else if(tc_eta->at(itc)>0. and genjetpt_posEta>=0.){
 	//if(hgcalInnerEta){
 	// totTCpt_posEta += tc_pt->at(itc);
 	// if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE += tc_pt->at(itc);
 	//}
-	//if(minRoz<tcRoZCut){
-	hTCPhiPos->Fill(TMath::RadToDeg()*tc_phi->at(itc));
-	totTCpt_posEta += tcpt;
-	if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE += tcpt;
-	//}
+	if(minRoz<tcRoZCut){
+	  totTCpt_posEta += tcpt;
+	  if(tc_layer->at(itc)<=26) totTCpt_posEta_CEE += tcpt;
+	}
 	tcXoZposEta += tcpt*tc_x->at(itc)/tc_z->at(itc) ;
 	tcYoZposEta += tcpt*tc_y->at(itc)/tc_z->at(itc) ;
 	tcPhi_posEta += tcpt*tc_phi->at(itc) ;
@@ -1504,35 +1301,18 @@ int main(int argc, char** argv)
 	  pTCPtvsGenPtByTCPtsum->Fill(tcptMeV, (clstPt/totTCpt_posEta_iter1));
 	  pTCPtvsGenPtByTCPtsumLogx->Fill(tcptMeV, (clstPt/totTCpt_posEta_iter1));
 	  pTCPtvsGenPtByTCPtsumDet[detType-1]->Fill(tcptMeV, (clstPt/totTCpt_posEta_iter1));
+	  hGenTCPtResovsTCPt1D[binPt-1]->Fill( (clstPt - totTCpt_posEta_iter1) );
+	  hGenTCPtResovsTCPt1DLog[binPtLog-1]->Fill( (clstPt - totTCpt_posEta_iter1) );
 	  hGenjetByTCPtSumVsTCPt1DLog[binPtLog-1]->Fill( clstPt/totTCpt_posEta_iter1 );
-     	  hGenjetByTCPtSumVsTCPt2DDetLog[detType-1][binPtLog-1]->Fill( clstPt/totTCpt_posEta_iter1 );
-	  hGenjetByTCPtSumVsTCPt2DLog[binEta-1][binPtLog-1]->Fill( clstPt/totTCpt_posEta_iter1 );
-	  //hGenjetByTCPtSumVsTCPt3DLog[detType-1][binEta-1][binPtLog-1]->Fill( clstPt/totTCpt_posEta_iter1 );
-	}
-	h2GenjetTCPtsumvsGenPt->Fill(clstPt, clstPt/totTCpt_posEta_iter1); 
-	h2GenjetTCPtsumvsGenEta->Fill(fabs(clstEta), clstPt/totTCpt_posEta_iter1);
-	pGenjetTCPtsumvsGenPt->Fill(clstPt, clstPt/totTCpt_posEta_iter1); 
-	pGenjetTCPtsumvsGenEta->Fill(fabs(clstEta), clstPt/totTCpt_posEta_iter1);
-	if(hgcalInnerEta){
-	  h2GenjetTCPtsumvsGenPt_CoreEta->Fill(clstPt, clstPt/totTCpt_posEta_CEE_core_iter1);
-	  pGenjetTCPtsumvsGenPt_CoreEta->Fill(clstPt, clstPt/totTCpt_posEta_CEE_core_iter1);
-	}
-	if(tc_layer->at(itc)<=26){
-	  h2GenjetTCPtsumvsGenPt_CEE->Fill(clstPt, clstPt/totTCpt_posEta_CEE_iter1);
-	  h2GenjetTCPtsumvsGenEta_CEE->Fill(fabs(clstEta), clstPt/totTCpt_posEta_CEE_iter1);
-	  pGenjetTCPtsumvsGenPt_CEE->Fill(clstPt, clstPt/totTCpt_posEta_CEE_iter1);
-	  pGenjetTCPtsumvsGenEta_CEE->Fill(fabs(clstEta), clstPt/totTCpt_posEta_CEE_iter1);
-	  if(hgcalInnerEta){
-	    h2GenjetTCPtsumvsGenPt_CEE_CoreEta->Fill(clstPt, clstPt/totTCpt_posEta_CEE_core_iter1);
-	    pGenjetTCPtsumvsGenPt_CEE_CoreEta->Fill(clstPt, clstPt/totTCpt_posEta_CEE_core_iter1);
-	    h2GenjetTCPtsumvsTCPtsum_CEE_CoreEta->Fill(totTCpt_posEta_CEE_core_iter1, clstPt/totTCpt_posEta_CEE_core_iter1);
-	    pGenjetTCPtsumvsTCPtsum_CEE_CoreEta->Fill(totTCpt_posEta_CEE_core_iter1, clstPt/totTCpt_posEta_CEE_core_iter1);
-	  }
+     	  hGenTCPtResovsTCPt2D[detType-1][binPt-1]->Fill( (clstPt - totTCpt_posEta_iter1) );
 	}
       }
       
+      tot_tc_pt += tcpt;
+      tot_tc_e += tc_energy->at(itc);
+      if(tc_layer->at(itc)==2) nofTCsTrigL2++;
+      if(tc_layer->at(itc)==9) nofTCsTrigL9++;
     }//end of TC loop
-    
     //rearrgdtcs.clear();
     if(doPrint){
       int nofTCTots = 0;
@@ -1546,63 +1326,37 @@ int main(int argc, char** argv)
 		<< std::endl;
       
     }
-    
-    for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
-      
-      int ijet = jetlist.at(ipjet).index ;
-
-      //if(fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore){
-      //if(fabs(genjet_eta->at(ijet))>1.6 and fabs(genjet_eta->at(ijet))<3.0){
-	if(genjet_eta->at(ijet)<0 and genjetpt_negEta>-1. and totTCpt_negEta_iter1>0.){	
-	  h2TCvsGenPt->Fill(genjetpt_negEta, totTCpt_negEta_iter1);
-	  pTCvsGenPt->Fill(genjetpt_negEta, totTCpt_negEta_iter1);
-	  h2TCvsGenPt_CEE->Fill(genjetpt_negEta, totTCpt_negEta_CEE_iter1);
-	  pTCvsGenPt_CEE->Fill(genjetpt_negEta, totTCpt_negEta_CEE_iter1);
-	}else if(genjet_eta->at(ijet)>0 and genjetpt_posEta>-1. and totTCpt_posEta_iter1>0.){
-	  h2TCvsGenPt->Fill(genjetpt_posEta, totTCpt_posEta_iter1);
-	  pTCvsGenPt->Fill(genjetpt_posEta, totTCpt_posEta_iter1);
-	  h2TCvsGenPt_CEE->Fill(genjetpt_posEta, totTCpt_posEta_CEE_iter1);
-	  pTCvsGenPt_CEE->Fill(genjetpt_posEta, totTCpt_posEta_CEE_iter1);
-	}
-	//}
-      
-      if(fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore){
-	if(genjet_eta->at(ijet)<0 and genjetpt_negEta>-1. and totTCpt_negEta_core_iter1>0.){
-	  h2TCvsGenPt_CoreEta->Fill(genjetpt_negEta, totTCpt_negEta_core_iter1);
-	  pTCvsGenPt_CoreEta->Fill(genjetpt_negEta, totTCpt_negEta_core_iter1);
-	  h2TCvsGenPt_CEE_CoreEta->Fill(genjetpt_negEta, totTCpt_negEta_CEE_core_iter1);
-	  pTCvsGenPt_CEE_CoreEta->Fill(genjetpt_negEta, totTCpt_negEta_CEE_core_iter1);
-	}else if(genjet_eta->at(ijet)>0 and genjetpt_posEta>-1. and totTCpt_posEta_core_iter1>0.){
-	  h2TCvsGenPt_CoreEta->Fill(genjetpt_posEta, totTCpt_posEta_core_iter1);
-	  pTCvsGenPt_CoreEta->Fill(genjetpt_posEta, totTCpt_posEta_core_iter1);
-	  h2TCvsGenPt_CEE_CoreEta->Fill(genjetpt_posEta, totTCpt_posEta_CEE_core_iter1);
-	  pTCvsGenPt_CEE_CoreEta->Fill(genjetpt_posEta, totTCpt_posEta_CEE_core_iter1);
-	}
+    if(fabs(maxTCPt_posEta_genEta)>1.6 and fabs(maxTCPt_posEta_genEta)<3.0){
+      hMaxLayerE->Fill(maxTCPt_posEta_layer,totTCpt_posEta);
+      hMaxDetE->Fill(maxTCPt_posEta_detType,totTCpt_posEta);
+      if(maxTCPt_posEta_layer<=26){
+	h2TCvsGenPt->Fill(maxTCPt_posEta, maxTCPt_posEta_genPt);
+	pTCvsGenPt->Fill(maxTCPt_posEta, maxTCPt_posEta_genPt);
+	h2GenvsTCPt->Fill(maxTCPt_posEta_genPt, maxTCPt_posEta);
+	pGenvsTCPt->Fill(maxTCPt_posEta_genPt, maxTCPt_posEta);
       }
-
-      if(fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore){
-	if(genjet_eta->at(ijet)<0 and genjetpt_negEta>-1. and totTCpt_negEta_core_rozcut0p15_iter1>0.){
-	  h2TCvsGenPt_CoreEta_rozcut0p15->Fill(genjetpt_negEta, totTCpt_negEta_core_rozcut0p15_iter1);
-	  pTCvsGenPt_CoreEta_rozcut0p15->Fill(genjetpt_negEta, totTCpt_negEta_core_rozcut0p15_iter1);
-	}else if(genjet_eta->at(ijet)>0 and genjetpt_posEta>-1. and totTCpt_posEta_core_rozcut0p15_iter1>0.){
-	  h2TCvsGenPt_CoreEta_rozcut0p15->Fill(genjetpt_posEta, totTCpt_posEta_core_rozcut0p15_iter1);
-	  pTCvsGenPt_CoreEta_rozcut0p15->Fill(genjetpt_posEta, totTCpt_posEta_core_rozcut0p15_iter1);
-	}
-      }
-      
-      if(fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore){
-	if(genjet_eta->at(ijet)<0 and genjetpt_negEta>-1. and totTCpt_negEta_core_mpptcut2_iter1>0.){
-	  h2TCvsGenPt_CoreEta_mpptcut2->Fill(genjetpt_negEta, totTCpt_negEta_core_mpptcut2_iter1);
-	  pTCvsGenPt_CoreEta_mpptcut2->Fill(genjetpt_negEta, totTCpt_negEta_core_mpptcut2_iter1);
-	}else if(genjet_eta->at(ijet)>0 and genjetpt_posEta>-1. and totTCpt_posEta_core_mpptcut2_iter1>0.){
-	  h2TCvsGenPt_CoreEta_mpptcut2->Fill(genjetpt_posEta, totTCpt_posEta_core_mpptcut2_iter1);
-	  pTCvsGenPt_CoreEta_mpptcut2->Fill(genjetpt_posEta, totTCpt_posEta_core_mpptcut2_iter1);
-	}
-      }
-      
-      
+      h2TCvsGenPtDet[maxTCPt_posEta_detType-1]->Fill(maxTCPt_posEta, maxTCPt_posEta_genPt);
+      pTCvsGenPtDet[maxTCPt_posEta_detType-1]->Fill(maxTCPt_posEta, maxTCPt_posEta_genPt);
     }
     
+    if(fabs(maxTCPt_negEta_genEta)>1.6 and fabs(maxTCPt_negEta_genEta)<3.0){
+      hMaxLayerE->Fill(maxTCPt_negEta_layer,totTCpt_negEta);
+      hMaxDetE->Fill(maxTCPt_negEta_detType, totTCpt_negEta);
+      if(maxTCPt_negEta_layer<=26){
+	h2TCvsGenPt->Fill(maxTCPt_negEta, maxTCPt_negEta_genPt);
+	pTCvsGenPt->Fill(maxTCPt_negEta, maxTCPt_negEta_genPt);
+	h2GenvsTCPt->Fill(maxTCPt_negEta_genPt, maxTCPt_negEta);
+	pGenvsTCPt->Fill(maxTCPt_negEta_genPt, maxTCPt_negEta);
+      }
+      h2TCvsGenPtDet[maxTCPt_negEta_detType-1]->Fill(maxTCPt_negEta, maxTCPt_negEta_genPt);
+      pTCvsGenPtDet[maxTCPt_negEta_detType-1]->Fill(maxTCPt_negEta, maxTCPt_negEta_genPt);
+    }
+    
+    //if(doPrint)
+    //std::cout<<"tot_tc_pt : "<< tot_tc_pt << std::endl;
+    for (uint32_t isect = 0 ; isect < 6 ; isect++ ){
+      //std::cout << "ievent : " << ievent << ", isect: " << isect << ", size : " << vTcw[isect].size() << std::endl;
+    }
     
     ///////////////////=========== Emulation ============= ///////////////////////
     //std::vector<TPGClusterData> vCld[6];
@@ -1613,28 +1367,22 @@ int main(int argc, char** argv)
     }
     ///////////////////=========== Emulation ============= ///////////////////////
 
-    int nofClus1GeVp = 0, nofClus1GeVm = 0;
-    int nofClus3GeVp = 0, nofClus3GeVm = 0;
-    int nofClus5GeVp = 0, nofClus5GeVm = 0;
-    int nofClus10GeVp = 0, nofClus10GeVm = 0;
+    int nofClus1GeV = 0;
+    int nofClus3GeV = 0;
+    int nofClus5GeV = 0;
+    int nofClus10GeV = 0;
     for (uint32_t isect = 0 ; isect < 6 ; isect++ ){
       //for(TPGCluster const& clf : vCld[isect]){
-      if(isect==0){
-	nofClus1GeVm = 0;
-	nofClus3GeVm = 0;
-	nofClus5GeVm = 0;
-	nofClus10GeVm = 0;
-      }
-      if(isect==3){
-	nofClus1GeVp = 0;
-	nofClus3GeVp = 0;
-	nofClus5GeVp = 0;
-	nofClus10GeVp = 0;
+      if(isect==0 or isect==3){
+	nofClus1GeV = 0;
+	nofClus3GeV = 0;
+	nofClus5GeV = 0;
+	nofClus10GeV = 0;
       }
       int iclus = 0;
       for(TPGCluster const& clf : vCld[isect]){
 	double tcPtSum = (clf.getGlobalEtaRad(isect) < 0) ? totTCpt_negEta : totTCpt_posEta ;
-	if(doPrint and clf.getEnergyGeV()>3.0){
+	if(doPrint){
 	  std::cout << "iclus-ievent: " << std::fixed << std::setprecision(2) << std::setw(4) << ievent
 		    << ", sector: " << std::setw(5) << isect
 		    << ", iclus: " << std::setw(5) << iclus++
@@ -1654,38 +1402,20 @@ int main(int argc, char** argv)
 		    << std::endl;
 	}
 
-	if(isect<=2){
-	  if(clf.getEnergyGeV()>1.0) nofClus1GeVm++;
-	  if(clf.getEnergyGeV()>3.0) nofClus3GeVm++;	
-	  if(clf.getEnergyGeV()>5.0) nofClus5GeVm++;
-	  if(clf.getEnergyGeV()>10.0) nofClus10GeVm++;
-	}else{
-	  if(clf.getEnergyGeV()>1.0) nofClus1GeVp++;
-	  if(clf.getEnergyGeV()>3.0) nofClus3GeVp++;	
-	  if(clf.getEnergyGeV()>5.0) nofClus5GeVp++;
-	  if(clf.getEnergyGeV()>10.0) nofClus10GeVp++;
-	}
+	if(clf.getEnergyGeV()>1.0) nofClus1GeV++;
+	if(clf.getEnergyGeV()>3.0) nofClus3GeV++;	
+	if(clf.getEnergyGeV()>5.0) nofClus5GeV++;
+	if(clf.getEnergyGeV()>10.0) nofClus10GeV++;
+
       }//cluster loop
     }//sector loop
-    if(nofClus1GeVm>0) hNClus1GeV->Fill(nofClus1GeVm);
-    if(nofClus1GeVp>0) hNClus1GeV->Fill(nofClus1GeVp);
-    if(nofClus3GeVm>0) hNClus3GeV->Fill(nofClus3GeVm);
-    if(nofClus3GeVp>0) hNClus3GeV->Fill(nofClus3GeVp);
-    if(nofClus5GeVm>0) hNClus5GeV->Fill(nofClus5GeVm);
-    if(nofClus5GeVp>0) hNClus5GeV->Fill(nofClus5GeVp);
-    if(nofClus10GeVm>0) hNClus10GeV->Fill(nofClus10GeVm);
-    if(nofClus10GeVp>0) hNClus10GeV->Fill(nofClus10GeVp);
     
     for(int ipjet=0; ipjet < jetlist.size() ; ipjet++ ){
       int ijet = jetlist.at(ipjet).index ;
       //hEt->Fill(genjet_pt->at(ijet));
-      double genjetpt = (genjet_eta->at(ijet) < 0) ? genjetpt_negEta : genjetpt_posEta;
-      if(genjetpt==-1.) continue;      
       int minsect = (genjet_eta->at(ijet) < 0) ? 0 : 3;
       int maxsect = (genjet_eta->at(ijet) < 0) ? 3 : 6;
       double tcPtSum = (genjet_eta->at(ijet) < 0) ? totTCpt_negEta : totTCpt_posEta ;
-      double tcPtSum_core = (genjet_eta->at(ijet) < 0) ? totTCpt_negEta_core_iter1 : totTCpt_posEta_core_iter1 ;
-      double tcPtSum_core_tcroz = (genjet_eta->at(ijet) < 0) ? totTCpt_negEta_core_rozcut0p15_iter1 : totTCpt_posEta_core_rozcut0p15_iter1 ;
       double tcPtSum_CEE = (genjet_eta->at(ijet) < 0) ? totTCpt_negEta_CEE : totTCpt_posEta_CEE ;
       int tcDetType = (genjet_eta->at(ijet) < 0) ? maxTCPt_negEta_detType : maxTCPt_posEta_detType ;
       double tcPt = (genjet_eta->at(ijet) < 0) ? maxTCPt_negEta : maxTCPt_posEta ;
@@ -1719,7 +1449,7 @@ int main(int argc, char** argv)
 	  totClusE += clf.getEnergyGeV() ; 
 	}//cluster loop
       }//sector loop
-      //closestClusE *= fClusCorr->Eval(closestClusE);
+      closestClusE *= fClusCorr->Eval(closestClusE);
       
       double minRoz_cl3d = 1., closestClusE_cl3d= 0., closestClusEta_cl3d= 0.;
       for(int icl3=0;icl3<cl3d_n;icl3++){	  
@@ -1733,7 +1463,7 @@ int main(int argc, char** argv)
 	  closestClusEta_cl3d = cl3d_eta->at(icl3);
 	}//minRoZ condition
       }//cluster loop
-      //closestClusE_cl3d *= fcl3d->Eval(closestClusE_cl3d);
+      closestClusE_cl3d *= fcl3d->Eval(closestClusE_cl3d);
       
       bool hasClosestFound = (minRoz<1.)?true:false;
       bool hasClosestFound_cl3d = (minRoz_cl3d<1.)?true:false;
@@ -1741,11 +1471,10 @@ int main(int argc, char** argv)
       // bool hasClosestFoundPtTot = (hasClosestFound and totClusE>pt_clusThresh)?true:false;
       //Eta : outermost(Sci) = 1.321, innermost(Si) =  3.152
       bool hgcalEta = (fabs(genjet_eta->at(ijet))>1.321 and fabs(genjet_eta->at(ijet))<3.152) ? true : false;
-      bool hgcalInnerEta = (fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore) ? true : false;
-      // bool hgcalOuterEta = (fabs(genjet_eta->at(ijet))>etaMinCore and fabs(genjet_eta->at(ijet))<etaMaxCore5) ? true : false;
+      bool hgcalInnerEta = (fabs(genjet_eta->at(ijet))>1.6 and fabs(genjet_eta->at(ijet))<3.0) ? true : false;
+      // bool hgcalOuterEta = (fabs(genjet_eta->at(ijet))>1.6 and fabs(genjet_eta->at(ijet))<3.05) ? true : false;
       
-      clusRoz->Fill(minRoz);
-      clusRozPtwt->Fill(minRoz, genjet_pt->at(ijet));
+      clusRoz->Fill(minRoz, genjet_pt->at(ijet));
       hJetEtaPtBin->Fill(genjet_eta->at(ijet), genjet_pt->at(ijet));
       GetHistoBin(hJetEtaPtBin,genjet_eta->at(ijet), genjet_pt->at(ijet), binEta, binPt);
       GetHistoBin(hJetEtaPtBin,closestClusEta, closestClusE, binEtaClus, binPtClus);
@@ -1754,66 +1483,66 @@ int main(int argc, char** argv)
       if(hgcalEta and hasClosestFound) {
 	hJetEtaPtBinDet->Fill( fabs(genjet_eta->at(ijet)), genjet_pt->at(ijet));
 	
-	hGenClusPtReso[binEtaClus-1][binPtClus-1]->Fill( genjet_pt->at(ijet)/closestClusE );
-	hGenTCPtReso[binEtaClus-1][binPtClus-1]->Fill( genjet_pt->at(ijet)/tcPtSum );
-	hTCClusPtReso[binEtaClus-1][binPtClus-1]->Fill( tcPtSum/closestClusE );
+	hGenClusPtReso[binEta-1][binPt-1]->Fill( (genjet_pt->at(ijet) - closestClusE) );
+	hGenTCPtReso[binEta-1][binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum_CEE) );
+	//hTCClusPtReso[binEta-1][binPt-1]->Fill( (tcPtSum - closestClusE) );
+	hTCClusPtReso[binEta-1][binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );
 	
-	hGenClusPtReso3D[tcDetType-1][binEtaClus-1][binPtClus-1]->Fill( genjet_pt->at(ijet)/closestClusE );
+	hGenClusPtReso3D[tcDetType-1][binEta-1][binPt-1]->Fill( (genjet_pt->at(ijet) - closestClusE) );
 	
 	if(tcDetType<=3){
-	  hTCElossCEEID[binEtaClus-1]->Fill( tcPtSum_CEE );
-	  hTCElossCE_E[binEtaClus-1][binPtClus-1]->Fill( tcPtSum_CEE );
-	  // hTCElossCE_E_3D[tcDetType-1][binEtaClus-1][binPtClus-1]->Fill( tcPtSum_CEE );
+	  hTCElossCEEID[binEta-1]->Fill( tcPtSum_CEE );
+	  hTCElossCE_E[binEta-1][binPt-1]->Fill( tcPtSum_CEE );
+	  // hTCElossCE_E_3D[tcDetType-1][binEta-1][binPt-1]->Fill( tcPtSum_CEE );
 	}else{
-	  hTCElossCEHID[binEtaClus-1]->Fill( (tcPtSum-tcPtSum_CEE) );
-	  hTCElossCE_H[binEtaClus-1][binPtClus-1]->Fill( (tcPtSum-tcPtSum_CEE) );
-	  // hTCElossCE_H_3D[tcDetType-1][binEtaClus-1][binPtClus-1]->Fill( (tcPtSum-tcPtSum_CEE) );
+	  hTCElossCEHID[binEta-1]->Fill( (tcPtSum-tcPtSum_CEE) );
+	  hTCElossCE_H[binEta-1][binPt-1]->Fill( (tcPtSum-tcPtSum_CEE) );
+	  // hTCElossCE_H_3D[tcDetType-1][binEta-1][binPt-1]->Fill( (tcPtSum-tcPtSum_CEE) );
 
-	  //hHadGenClusPtReso2D[binEtaClus-1][binPtClus-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
-	  // hHadGenClusPtReso3D[tcDetType-1][binEtaClus-1][binPtClus-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
+	  //hHadGenClusPtReso2D[binEta-1][binPt-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
+	  // hHadGenClusPtReso3D[tcDetType-1][binEta-1][binPt-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
 	}
-
-	h2ClusVsTCPt->Fill(tcPtSum, closestClusE);
-	h2TotClusVsTCPt->Fill(tcPtSum, totClusE);
-	if(totClusE>tcPtSum){
-	  std::cerr<<"Issue: totClusE>tcPtSum in file : " << fin->GetName() << ", Event: " << ievent << ", ijet: " << ijet << ", genjetpt: " << genjetpt << ", genjeteta: " << genjet_eta->at(ijet) << ", phi: " << genjet_phi->at(ijet) << ", tcPtSum: " << tcPtSum << ", totClusE:" << totClusE  << std::endl;
-	}
+	
 	if(hgcalInnerEta){
+	  // hGenClusPtReso1D[binPt-1]->Fill( (genjet_pt->at(ijet) - closestClusE) );
+	  // hGenTCPtReso1D[binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum_CEE) );
+	  // hTCClusPtReso1D[binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );	  
 	  
 	  hGenClusPtReso1D[binPtClus-1]->Fill( genjet_pt->at(ijet)/closestClusE );
 	  hGenTCPtReso1D[binPtClus-1]->Fill( genjet_pt->at(ijet)/tcPtSum );
 	  hTCClusPtReso1D[binPtClus-1]->Fill( tcPtSum/closestClusE );	  
 	  
-	  h2GenVsClusPt->Fill(closestClusE, genjet_pt->at(ijet));
-	  h2GenVsTCPt->Fill(tcPtSum, genjet_pt->at(ijet));
-	  h2TCVsClusPt->Fill( closestClusE, tcPtSum);
+	  h2GenVsClusPt->Fill(genjet_pt->at(ijet), closestClusE);
+	  h2GenVsTCPt->Fill(genjet_pt->at(ijet), tcPtSum);
+	  h2TCVsClusPt->Fill( tcPtSum, closestClusE);
 	  
-	  h2ClusVsTCPt_CoreEta->Fill(tcPtSum_core, closestClusE);
-	  h2ClusVsTCPt_CoreEta_rozcut0p15->Fill(tcPtSum_core_tcroz, closestClusE);
-
-	  h2TotClusVsTCPt_CoreEta->Fill(tcPtSum_core, totClusE);
-	  h2TotClusVsTCPt_CoreEta_rozcut0p15->Fill(tcPtSum_core_tcroz, totClusE);
-
-	  // GetHistoBin(hJetEtaTCPtBin,genjet_eta->at(ijet), tcPt, binEta, binPt);
-	  // hGenTCPtResovsTCPt1D[binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );
-	  // hGenTCPtResovsTCPt2D[tcDetType-1][binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );
+	  GetHistoBin(hJetEtaTCPtBin,genjet_eta->at(ijet), tcPt, binEta, binPt);
+	  hGenTCPtResovsTCPt1D[binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );
+	  hGenTCPtResovsTCPt2D[tcDetType-1][binPt-1]->Fill( (genjet_pt->at(ijet) - tcPtSum) );
+	  
+	  // if(tcDetType>=4)  hHadGenClusPtReso1D[binPt-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
+	  // hHadGenTCPtReso1D[binPt-1]->Fill( ((genjet_pt->at(ijet)-tcPtSum_CEE) - closestClusE) );
+	  // hHadTCClusPtReso1D[binPt-1]->Fill( ((tcPtSum-tcPtSum_CEE) - closestClusE) );
+	  // hHadGenClusPtReso1D[binPt-1]->Fill((genjet_pt->at(ijet)-tcPtSum_CEE));
+	  // hHadGenTCPtReso1D[binPt-1]->Fill( tcPtSum - tcPtSum_CEE );
+	  // hHadTCClusPtReso1D[binPt-1]->Fill( tcPtSum_CEE );
 	  
 	  if(hasClosestFound_cl3d){
 	    hNewOldPtReso1Dcl3d[binPtClus-1]->Fill( closestClusE/closestClusE_cl3d );
-	    h2NewVsOldPtcl3d->Fill( closestClusE, closestClusE_cl3d);
-	    pNewVsOldPtcl3d->Fill( closestClusE, closestClusE_cl3d);
+	    h2NewVsOldPtcl3d->Fill( closestClusE_cl3d , closestClusE);
+	    pNewVsOldPtcl3d->Fill( closestClusE_cl3d , closestClusE);
 	  }
 
 	}
       }
       if(hgcalInnerEta and hasClosestFound_cl3d) {
 	hGenClusPtReso1Dcl3d[binPt_cl3d-1]->Fill( genjet_pt->at(ijet)/closestClusE_cl3d );
-	h2GenVsClusPtcl3d->Fill(closestClusE_cl3d, genjet_pt->at(ijet));
+	h2GenVsClusPtcl3d->Fill(genjet_pt->at(ijet),closestClusE_cl3d);
       }
       
     }//jet loop 
     
-    
+
     
     for(uint16_t i=0;i<6;i++) {
       vCld[i].clear();
@@ -1850,7 +1579,6 @@ int main(int argc, char** argv)
     genjet_phi->clear();
     genjet_energy->clear();
     tc_pt->clear();
-    tc_mipPt->clear();
     tc_eta->clear();
     tc_phi->clear();
     tc_layer->clear();
@@ -1872,9 +1600,6 @@ int main(int argc, char** argv)
   TH1D *tcRoz_cumul = (TH1D *)tcRoz->GetCumulative(kTRUE,"_cumul");
   tcRoz_cumul->Scale(1./tcRoz_cumul->GetBinContent(tcRoz_cumul->GetMaximumBin()));
 
-  TH1D *tcRozPtwt_cumul = (TH1D *)tcRozPtwt->GetCumulative(kTRUE,"_cumul");
-  tcRozPtwt_cumul->Scale(1./tcRozPtwt_cumul->GetBinContent(tcRozPtwt_cumul->GetMaximumBin()));
-
   TH1D *clusRoz_cumul = (TH1D *)clusRoz->GetCumulative(kTRUE,"_cumul");
   clusRoz_cumul->Scale(1./clusRoz_cumul->GetBinContent(clusRoz_cumul->GetMaximumBin()));
 
@@ -1882,95 +1607,29 @@ int main(int argc, char** argv)
   hGenvsJetPtDiff->Write();
   pGenvsJetPtDiff->Write();
   hGenvsJetDR->Write();
-  pPtDiffvsDR->Write();
-  h2PtDiffvsDR->Write();
   hJetEtaPtBin->Write();
   hJetEtaPtBinDet->Write();
+  h2TCvsGenPt->Write();
+  pTCvsGenPt->Write();
+  h2GenvsTCPt->Write();
+  pGenvsTCPt->Write();
   h2TCPtvsGenPtByTCPtsum->Write();
   pTCPtvsGenPtByTCPtsum->Write();
   pTCPtvsGenPtByTCPtsumLogx->Write();
   tcRoz->Write();
-  htcMipPt->Write();
-  tcRozCore->Write();
   tcRoz_cumul->Write();
-  tcRozPtwt->Write();
-  tcRozPtwt_cumul->Write();
   clusRoz->Write();
   clusRoz_cumul->Write();
-  clusRozPtwt->Write();
   hNJets->Write();
   hNJetsF1->Write();
-  hNClus1GeV->Write();
-  hNClus3GeV->Write();
-  hNClus5GeV->Write();
-  hNClus10GeV->Write();
-  //=============================================
-  TDirectory *dir = fout->mkdir("TC_Scale");
-  dir->cd();
-  //=
-  h2TCvsGenPt->Write();
-  h2TCvsGenPt_CEE->Write();
-  h2TCvsGenPt_CEE_CoreEta->Write();
-  pTCvsGenPt->Write();
-  pTCvsGenPt_CEE->Write();
-  pTCvsGenPt_CEE_CoreEta->Write();  
-  //=
-  h2TCvsGenPt_CoreEta->Write();
-  pTCvsGenPt_CoreEta->Write();
-  h2TCvsGenPt_CoreEta_rozcut0p15->Write();
-  pTCvsGenPt_CoreEta_rozcut0p15->Write();
-  h2TCvsGenPt_CoreEta_mpptcut2->Write();
-  pTCvsGenPt_CoreEta_mpptcut2->Write();
-  //=
-  h2GenjetTCPtsumvsGenPt->Write();
-  h2GenjetTCPtsumvsGenEta->Write();
-  pGenjetTCPtsumvsGenPt->Write();
-  pGenjetTCPtsumvsGenEta->Write();
-  //=
-  h2GenjetTCPtsumvsGenPt_CEE->Write();
-  h2GenjetTCPtsumvsGenEta_CEE->Write();
-  h2GenjetTCPtsumvsGenPt_CEE_CoreEta->Write();
-  pGenjetTCPtsumvsGenPt_CEE->Write();
-  pGenjetTCPtsumvsGenEta_CEE->Write();
-  pGenjetTCPtsumvsGenPt_CEE_CoreEta->Write();
-  //=
-  h2GenjetTCPtsumvsTCPtsum_CEE_CoreEta->Write();
-  pGenjetTCPtsumvsTCPtsum_CEE_CoreEta->Write();
-  //=
-  h2GenjetTCPtsumvsGenPt_CoreEta->Write();
-  pGenjetTCPtsumvsGenPt_CoreEta->Write();
-  //=
-  fout->cd();
-  //=============================================
-  dir = fout->mkdir("TC_histos");
-  dir->cd();
-  hTCPt->Write();
-  hTCEta->Write();
-  hTCPhiPos->Write();
-  hTCPhiNeg->Write();
-  h2TCEtaPhi->Write();
-  fout->cd();
-  //=============================================
-  dir = fout->mkdir("Clus");
-  dir->cd();
   h2GenVsClusPt->Write();
   h2GenVsTCPt->Write();
   h2TCVsClusPt->Write();
-  //=
   h2GenVsClusPtcl3d->Write();  
   h2NewVsOldPtcl3d->Write();  
-  pNewVsOldPtcl3d->Write();
-  //=
-  h2ClusVsTCPt->Write();  
-  h2ClusVsTCPt_CoreEta->Write();  
-  h2ClusVsTCPt_CoreEta_rozcut0p15->Write();
-  h2TotClusVsTCPt->Write();  
-  h2TotClusVsTCPt_CoreEta->Write();  
-  h2TotClusVsTCPt_CoreEta_rozcut0p15->Write();  
-  //=
-  fout->cd();
+  pNewVsOldPtcl3d->Write();  
   //=============================================
-  dir = fout->mkdir("Reso_1D");
+  TDirectory *dir = fout->mkdir("Reso_1D");
   dir->cd();
   for(int ipt=0;ipt<nJetPtBins;ipt++) hGenClusPtReso1D[ipt]->Write();
   for(int ipt=0;ipt<nJetPtBins;ipt++) hGenTCPtReso1D[ipt]->Write();
@@ -1982,9 +1641,12 @@ int main(int argc, char** argv)
   for(int ipt=0;ipt<nJetPtBins;ipt++) hNewOldPtReso1Dcl3d[ipt]->Write();
   for(int ieta=0;ieta<nJetEtaBins;ieta++) hTCElossCEEID[ieta]->Write();
   for(int ieta=0;ieta<nJetEtaBins;ieta++) hTCElossCEHID[ieta]->Write();
+  for(int idet=0;idet<7;idet++) h2TCvsGenPtDet[idet]->Write();
+  for(int idet=0;idet<7;idet++) pTCvsGenPtDet[idet]->Write();
+  for(int ipt=0;ipt<nTCPtBins;ipt++) hGenTCPtResovsTCPt1D[ipt]->Write();
+  for(int ipt=0;ipt<int(logbins.size()-1);ipt++) hGenTCPtResovsTCPt1DLog[ipt]->Write();
   for(int ipt=0;ipt<int(logbins.size()-1);ipt++) hGenjetByTCPtSumVsTCPt1DLog[ipt]->Write();
   for(int idet=0;idet<7;idet++) pTCPtvsGenPtByTCPtsumDet[idet]->Write();
-  for(int idet=0;idet<7;idet++) tcRozDet[idet]->Write();
   fout->cd();
   //=============================================
   dir = fout->mkdir("Reso_2D");
@@ -2020,17 +1682,26 @@ int main(int argc, char** argv)
   }//jet eta
 
   for(int idet=0;idet<7;idet++)
-    for(int ipt=0;ipt<int(logbins.size()-1);ipt++)
-      hGenjetByTCPtSumVsTCPt2DDetLog[idet][ipt]->Write();
-
-  for(int ieta=0;ieta<nJetEtaBins;ieta++)
-    for(int ipt=0;ipt<int(logbins.size()-1);ipt++)
-      hGenjetByTCPtSumVsTCPt2DLog[ieta][ipt]->Write();
+  for(int ipt=0;ipt<nTCPtBins;ipt++)
+    hGenTCPtResovsTCPt2D[idet][ipt]->Write();
+  // for(int ieta=0;ieta<nJetEtaBins;ieta++){
+  //   for(int ipt=0;ipt<nJetPtBins;ipt++){
+  //     hHadGenClusPtReso2D[ieta][ipt]->Write();
+  //   }// jet et
+  // }//jet eta
   
   fout->cd();
   //=============================================
   dir = fout->mkdir("Reso_3D");
   dir->cd();
+  // for(int idet=0;idet<7;idet++){
+  //   for(int ieta=0;ieta<nJetEtaBins;ieta++){
+  //     for(int ipt=0;ipt<nJetPtBins;ipt++){
+  // 	hHadGenClusPtReso3D[idet][ieta][ipt]->Write();
+  //     }// jet et
+  //   }//jet eta
+  // }//idet
+
   for(int idet=0;idet<7;idet++){
     for(int ieta=0;ieta<nJetEtaBins;ieta++){
       for(int ipt=0;ipt<nJetPtBins;ipt++){
@@ -2038,22 +1709,16 @@ int main(int argc, char** argv)
       }// jet et
     }//jet eta
   }//idet
-
-
-  fout->cd();
-  // //=============================================
-
-  // //=============================================
-  // dir = fout->mkdir("Reso_3D_TC");
-  // dir->cd();
+  
   // for(int idet=0;idet<7;idet++){
   //   for(int ieta=0;ieta<nJetEtaBins;ieta++){
-  //     for(int ipt=0;ipt<int(logbins.size()-1);ipt++){
-  // 	hGenjetByTCPtSumVsTCPt3DLog[idet][ieta][ipt]->Write();
+  //     for(int ipt=0;ipt<nJetPtBins;ipt++){
+  // 	hGenTCPtReso3D[idet][ieta][ipt]->Write();
   //     }// jet et
   //   }//jet eta
   // }//idet
-  // fout->cd();
+  
+  fout->cd();
   // //=============================================
   // dir = fout->mkdir("CEE_3D");
   // dir->cd();
@@ -2077,6 +1742,8 @@ int main(int argc, char** argv)
   // }//idet  
   // fout->cd();
   // //=============================================
+  hMaxLayerE->Write();
+  hMaxDetE->Write();
   
   fout->Close();
   delete fout;
